% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addFeature}
\alias{addFeature}
\title{Add a new feature}
\usage{
addFeature(object, feature_name, feature_df, overwrite = FALSE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{feature_name}{The name of the feature that is to be added
to the obejct specified as a single character value.}

\item{feature_df}{A data.frame that contains the variables
\emph{barcodes, samples, \code{feature_name}}.}

\item{overwrite}{Logical. If the specified feature name already exists in the
current spata-object this argument must be set to TRUE in order to overwrite it.}
}
\value{
An updated spata-object.
}
\description{
Adds a new variable to the objects feature data.
}
\details{
Eventually the new feature will be joined via \code{dplyr::left_join()} over the
key-variable \emph{barcodes}. Additional steps secure the joining process.
}
