% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addFeatures}
\alias{addFeatures}
\alias{discardFeatures}
\title{Add a new feature}
\usage{
addFeatures(
  object,
  feature_names,
  feature_df,
  key_variable = "barcodes",
  overwrite = FALSE,
  of_sample = ""
)

discardFeatures(object, feature_names)
}
\arguments{
\item{object}{A valid spata-object.}

\item{feature_df}{A data.frame that contains the feature and the key-variables.}

\item{key_variable}{Character value. Either \emph{'barcodes'} or \emph{'coordinates'}.
If set to \emph{'coordinates'} the \code{feature_df}-input must contain numeric x- and
y- variables.

Key variables are variables in a data.frame that uniquely identify each observation -
in this case each barcode-spot. In SPATA the barcode-variable is a key-variable on its own,
x- and y-coordinates work as key-variables if they are used combined.}

\item{overwrite}{Logical. If the specified feature name already exists in the
current spata-object this argument must be set to TRUE in order to overwrite it.}
}
\value{
An updated spata-object.
}
\description{
Adds a new variable to the objects feature data.
}
\details{
Eventually the new feature will be joined via \code{dplyr::left_join()} over the
key-variables \emph{barcodes} or \emph{x} and \emph{y}. Additional steps secure the joining process.
}
