% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{addGeneMetaData}
\alias{addGeneMetaData}
\title{Add gene meta data to the object}
\usage{
addGeneMetaData(object, of_sample = "", meta_data_list)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{meta_data_list}{Output list of \code{computeGeneMetaData2()}. An additional
slot named \emph{mtr_name} needs to be added manually.}
}
\value{
An updated spata-object.
}
\description{
Safely adds the output of \code{computeGeneMetaData2()}
to the spata-object.
}
