% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R, R/add_family.R
\name{addGeneSet}
\alias{addGeneSet}
\title{Add a new gene set}
\usage{
addGeneSet(object, class_name, gs_name, genes, overwrite = FALSE)

addGeneSet(object, class_name, gs_name, genes, overwrite = FALSE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{class_name}{The class the gene set belongs to specified as a single character value.}

\item{gs_name}{The name of the new gene set specified as a single character value.}

\item{genes}{The genes of interest specified as a character vector.}

\item{overwrite}{Logical. Overwrites existing gene sets with the same \code{class_name} -
\code{gs_name} combination.}
}
\value{
An updated spata-object.

An updated spata-object.
}
\description{
Stores a new gene set in the spata-object.

Stores a new gene set in the spata-object.
}
\details{
Combines \code{class_name} and \code{gs_name} to the final gene set name.
Gene set classes and gene set names are separated by '_' and handled like this
in all additional gene set related functions which is why \code{class_name} must
not contain any '_'.

Combines \code{class_name} and \code{gs_name} to the final gene set name.
Gene set classes and gene set names are separated by '_' and handled like this
in all additional gene set related functions which is why \code{class_name} must
not contain any '_'.
}
