% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addUmapData}
\alias{addUmapData}
\alias{addTsneData}
\title{Add dimensional reductions}
\usage{
addUmapData(object, umap_df, overwrite = FALSE)

addTsneData(object, tsne_df, overwrite = FALSE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{umap_df}{A data.frame containing the character variables \emph{barcodes, sample} and
the numeric variables \emph{umap1, umap2}.}

\item{overwrite}{Logical. Must be set to TRUE in order to overwrite already existing data.}

\item{tsne_df}{A data.frame containing the character variables \emph{barcodes, sample} and
the numeric variables \emph{tsne1, tsne2}.}
}
\value{
An updated spata-object.
}
\description{
Adds or replaces dimensional reduction data. If the object contains
several sample the sample-variable of the input data.frame denotes the sample
belonging.
}
