% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-and-gene-sets.R
\name{adjustDefaultInstructions}
\alias{adjustDefaultInstructions}
\alias{adjustDirectoryInstructions}
\title{Adjust default instructions}
\usage{
adjustDefaultInstructions(object, ...)

adjustDirectoryInstructions(object, to, new_directory, combine_with_wd = "/")
}
\arguments{
\item{object}{A valid spata-object.}

\item{...}{Named arguments whoose default input you want to override.}

\item{new_directory}{Character value. The new directory under which
to store the object of interest. Overwrites the stored default directory.
Use \code{getDefaultDirectory()} to obtain the current set up.}

\item{combine_with_wd}{Character value or FALSE. If specified with a
character value (default: \emph{'/'}) the input of \code{new_directory}
is considered to be a relative directory and is combined with the
current working directory (\code{base::getwd()}) with the character string
specified as the separator. If set to FALSE the input of \code{new_directory}
is taken as is.}
}
\value{
An updated spata object.
}
\description{
Adjust default instructions
}
\examples{

 # Not run

 object <- adjustDefaultInstructions(object, pt_size = 4, smooth = FALSE)
}
