% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{assess_intra_cluster_distance}
\alias{assess_intra_cluster_distance}
\title{Title}
\usage{
assess_intra_cluster_distance(clustered_df)
}
\arguments{
\item{clustered_df}{dropped_df that contains an additional
cluster variable indicating the cluster belonging of
every barcode spot.}
}
\description{
Assesses the gene's tendency to cluster together.
The lower the \emph{intra_cluster_dist} the higher the tendency
to cluter.
}
\details{
The distances between all barcode spots of each cluster are
computed and normalized for the size of the cluster.
}
