% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-denoising.R
\name{assessAutoencoderOptions}
\alias{assessAutoencoderOptions}
\alias{assessAutoencoderOptions2}
\title{Assessment of Neural Network Set Up}
\usage{
assessAutoencoderOptions(
  object,
  activations,
  bottlenecks,
  layers = c(128, 64, 32),
  dropout = 0.1,
  epochs = 20,
  verbose = TRUE
)

assessAutoencoderOptions2(
  expr_mtr,
  activations,
  bottlenecks,
  layers = c(128, 64, 32),
  dropout = 0.1,
  epochs = 20,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{activations}{Character vector. Denotes the activation functions to be assessed.}

\item{bottlenecks}{Numeric vector. Denotes the different numbers of bottleneck neurons to be assessed.}

\item{layers}{Numeric vector of length 3. Denotes the number of neurons in the three hidden layers.
(default = c(128, 64, 32))}

\item{dropout}{Numeric value. Denotes the dropout. (default = 0.1)}

\item{epochs}{Numeric value. Denotes the epochs of the neural network. (default = 20)}

\item{expr_mtr}{The expression matrix that is to be used as input for the neural network.}
}
\value{
The spata object containing the list that holds the total variance measured by \code{irlba::prcomp_irlba()} after each
combination of activations/bottlenecks as well as the additional set up. If \code{assessAutoencoderOptions2} is called only
the list is returned.
}
\description{
Assesses different neural network set ups regarding
the activation function and the number of bottleneck neurons.
}
