% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autoencoder.R
\name{AutoEncoderDenoising}
\alias{AutoEncoderDenoising}
\title{Deep Autoencoder for Denoising}
\usage{
AutoEncoderDenoising(
  object,
  Layer = NULL,
  Bottleneck = NULL,
  Dropout = NULL,
  activation = NULL,
  epochs = NULL,
  val_plot = F
)
}
\arguments{
\item{Layer}{Set Number of Neurons in your 3 hidden layers (defauled = c("128", "64", "32"))}

\item{Bottleneck}{Set numbers of Bottleneck Neurons (defauled = "32")}

\item{Dropout}{Define Dropout (defauled = "0.1")}

\item{activation}{Set activation function (defauled = "relu")}

\item{epochs}{epochs of your NN (defauled = 20)}

\item{val_plot}{}
}
\value{
A spata-object.
}
\description{
This Function will return a SPATA object containing denoised data and re-analysis
}
