% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{average_genes}
\alias{average_genes}
\title{average_genes}
\usage{
average_genes(average_genes)
}
\arguments{
\item{average_genes}{Logical value. If set to TRUE the mean expression
of all specified genes will be calculated.}
}
\description{
average_genes
}
