% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern-analysis.R
\name{calculateDistance}
\alias{calculateDistance}
\alias{calculateDistanceLineplot}
\alias{calculateDistanceHeatmap}
\title{Spatial distance computation}
\usage{
calculateDistance(
  object,
  of_sample = "",
  variable,
  plot_type,
  method_gs = "mean",
  bins = 10,
  verbose = TRUE
)

calculateDistanceLineplot(
  object,
  of_sample = "",
  variable,
  method_gs = "mean",
  bins = 10,
  verbose = TRUE
)

calculateDistanceHeatmap(
  object,
  of_sample = "",
  variable,
  method_gs = "mean",
  bins = 10,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{plot_type}{Character value. Either \emph{'heatmap'} or \emph{'lineplot'}.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{bins}{Numeric value. If \code{variable} is a numeric variable
it specifies the number of bins in which the the values are divided.
Given to \code{dplyr::ntile()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A named list with four slots:

\itemize{
 \item{\emph{plot}: The lineplot or the heatmap.}
 \item{\emph{data}: The computed distance data.}
 \item{\emph{variable}: The variable specified.}
 \item{\emph{sample}: The sample specified.}}
}
\description{
- -
}
