% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_family.R
\name{check_assign}
\alias{check_assign}
\title{Check input}
\usage{
check_assign(assign, assign_name)
}
\arguments{
\item{assign}{Logical.}

\item{assign_name}{Character value.}
}
\description{
Make sure that no objects are overwritten.
}
