% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_family.R
\name{check_color_to}
\alias{check_color_to}
\title{Check input}
\usage{
check_color_to(
  color_to,
  all_features = character(),
  all_gene_sets = character(),
  all_genes = character()
)
}
\arguments{
\item{color_to}{character or list}

\item{all_features}{character. all features}

\item{all_gene_sets}{character. all gene sets}

\item{all_genes}{character. all genes}
}
\value{
a named list
}
\description{
Check input
}
