% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_family.R
\name{check_coordinates}
\alias{check_coordinates}
\title{Title}
\usage{
check_coordinates(data, x = "x", y = "y")
}
\arguments{
\item{data}{A data.frame containing numeric variables for x and y axis}

\item{x}{The name of the numeric variable to be plotted on the x axis.}

\item{y}{The name of the numeric variable to be plotted on the y axis.}
}
\description{
Title
}
\examples{

}
