% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_family.R
\name{check_coords_df}
\alias{check_coords_df}
\title{Check input}
\usage{
check_coords_df(coords_df)
}
\arguments{
\item{coords_df}{A data.frame that contains two character variables: \emph{
barcodes & sample}}
}
\value{
- If the input is just fine: the input.
- If the input suffices for the function to work but needs slight moderation:
a modified version of the input along with a warning-message.
- If the input does not suffice for the function to work: an error
}
\description{
The \code{check_()}-function family checks a functions input to
make sure that the function can run bugfree. If the input does not suffice a
helping message is printed for the user to know what needs to be adjusted in
order to make the function work.
}
