% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_gene_sets}
\alias{check_gene_sets}
\title{Check feature variables input}
\usage{
check_gene_sets(object, gene_sets, max_length = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{gene_sets}{The gene sets of interest specified as a character vector.}

\item{max_length}{The maximum number of features allowed.}
}
\value{
The original input, an adjusted version of it or an error. In the latter two
cases they print an informative message about what was going on.
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a character
vector of feature names, checks who of the features exist and checks additionally
if these features match the class requirements of \code{valid_classes}.

Returns an adjusted featurs-vector or raises an error.
}
