% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_genes}
\alias{check_genes}
\title{Check feature variables input}
\usage{
check_genes(object, genes, rna_assay = NULL, max_length = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{genes}{The genes of interest specified as a character vector.}

\item{rna_assay}{A rna-assay (e.g. derived from \code{exprMtr()}).}

\item{max_length}{The maximum number of features allowed.}
}
\value{
The original input, an adjusted version of it or an error. In the latter two
cases they print an informative message about what was going on.
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a character
vector of feature names, checks who of the features exist and checks additionally
if these features match the class requirements of \code{valid_classes}.

Returns an adjusted featurs-vector or raises an error.
}
