% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_method}
\alias{check_method}
\title{Check method input}
\usage{
check_method(
  method_dr = NULL,
  method_gs = NULL,
  method_de = NULL,
  method_padj = NULL
)
}
\arguments{
\item{method_dr}{Character value. The dimensional reduction method of
interest specified as a single character value. (Currently
either \emph{'UMAP'} or \emph{'TSNE'}).}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{method_de}{Character value. Given to argument \code{test.use} of \code{Seurat::FindAllMarkers()}.}

\item{method_padj}{Character value. The method according to which the adjusted p-values will
be calculated. Given to \code{stats::p.adjust()}.}
}
\value{
A logical value TRUE if nothing invalid has been detected or an informative
error message.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
