% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_pattern}
\alias{check_pattern}
\title{Title}
\usage{
check_pattern(object, patterns = "", method_pr = "hotspot", of_sample = NA)
}
\arguments{
\item{of_pattern}{}
}
\value{

}
\description{
Title
}
