% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_family.R
\name{check_pt_input}
\alias{check_pt_input}
\title{Check input}
\usage{
check_pt_input(pt_size = NULL, pt_alpha = NULL, pt_clrsp = NULL)
}
\arguments{
\item{pt_size}{The pt_size-input.}

\item{pt_alpha}{The pt_alpha-input.}

\item{pt_clrsp}{The pt_clrsp-input.}
}
\value{
- If the input is just fine: the input.
- If the input suffices for the function to work but needs slight moderation:
a modified version of the input along with a warning-message.
- If the input does not suffice for the function to work: an error
}
\description{
The \code{check_()}-function family checks a functions input to
make sure that the function can run bug-free. If the input does not suffice a
helping message is printed for the user to know what needs to be adjusted in
order to make the function work.
}
