% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_sample}
\alias{check_sample}
\title{Check feature variables input}
\usage{
check_sample(object, of_sample, desired_length = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}

\item{desired_length}{The length the input must have.}
}
\value{
The original input, an adjusted version of it or an error. In the latter two
cases they print an informative message about what was going on.
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a character
vector of feature names, checks who of the features exist and checks additionally
if these features match the class requirements of \code{valid_classes}.

Returns an adjusted featurs-vector or raises an error.
}
