% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_family.R
\name{check_smooth}
\alias{check_smooth}
\title{Check input}
\usage{
check_smooth(df, smooth, verbose = TRUE)
}
\arguments{
\item{df}{The data.frame that is to be smoothed.}

\item{smooth}{The smooth input.}

\item{verbose}{Logical.}
}
\value{
- If the input is just fine: the input.
- If the input suffices for the function to work but needs slight moderation:
a modified version of the input along with a warning-message.
- If the input does not suffice for the function to work: an error
}
\description{
Checks among other things whether \code{df} is spatially 'smoothable' -
whether it contains x and y variables.
}
