% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{clrp}
\alias{clrp}
\title{clrp}
\usage{
clrp(clrp)
}
\arguments{
\item{clpr}{Character value. The color panel to be used. Run \code{all_colorpanels()} to see
all valid input options.}
}
\description{
clrp
}
