% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{compute_pam_with_k_optimum}
\alias{compute_pam_with_k_optimum}
\title{pam-wrapper}
\usage{
compute_pam_with_k_optimum(dropped_df, k, ...)
}
\arguments{
\item{dropped_df}{A dropped data.frame.}

\item{k}{Numeric value. Denotes the number of clusters.}

\item{...}{Additional arguments given to \code{cluster::pam()}.}
}
\description{
pam-wrapper
}
