% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_family.R
\name{coordsSpatial}
\alias{coordsSpatial}
\alias{coordsTrajectory}
\alias{coordsDimRed}
\title{Extract coordinates or positions}
\usage{
coordsSpatial(object, of_sample, of_segment = NULL)

coordsTrajectory(object, of_sample, of_trajectory)

coordsDimRed(object, of_sample, method_dr = c("UMAP", "TSNE"))
}
\arguments{
\item{object}{A valid object of class \emph{spata}.}

\item{of_sample}{The sample(s) name specified as a character.}

\item{of_segment}{The segment name specified as a character of length one.}

\item{of_trajectory}{The trajectory name specified as a character of length one.}

\item{method_dr}{The dimensional reduction method from which to extract the
positions specified as a character of length one.}
}
\value{
A data.frame that contains in all cases the unique identifiers
(keys): \emph{barcodes, sample}.

Depending on the function called additional variables such as \emph{x, y,
umap1, umap2, tsne1 or tsne2}.
}
\description{
Extract coordinates or positions
}
