% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_family.R
\name{createSegmentation}
\alias{createSegmentation}
\title{Spatial Segmentation}
\usage{
createSegmentation(object = NULL)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
An updated version of the spata-object specified as \code{object}
now containing the information about all drawn segments.
}
\description{
The function \code{createSegmentation()} provides access to an
interactive mini-shiny application that allows to separate a sample into
several segments.
}
