% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{createSpataObject_10X}
\alias{createSpataObject_10X}
\title{Create a spata-object}
\usage{
createSpataObject_10X(
  input_paths,
  output_path,
  gene_set_path = NULL,
  sample_names,
  object_name,
  verbose = TRUE
)
}
\arguments{
\item{input_paths}{Character vector. Specifies the 10X visium-folders from
which to load the information. These folders must contain the following sub directories:

\itemize{
 \item{\emph{'/outs/filtered_feature_bc_matrix.h5'}}
 \item{\emph{'/outs/spatial/*.jpg}}
 }}

\item{output_path}{Character value. Specifies the folder in which to store
the object.}

\item{gene_set_path}{Character value (or NULL). Specifies the path to a
.RDS-file containing a data.frame that is to be used as input for slot @used_genesets.

 Must have the character-variables

   \itemize{
    \item{\emph{'ont'}: The gene set names.}
    \item{\emph{'gene'}: The belonging gene names.}
    }}

\item{sample_names}{Character vector. The sample names with which to refer to the
respective data. Must be of the same length as \code{input_paths}.}

\item{object_name}{Character value. The name-suffix for the file name under which the
spata-object is stored. Is prefixed with \emph{'spata-obj-'}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A spata-object.
}
\description{
Creates, saves and returns an object of class spata
from scratch.
}
