% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_family.R
\name{createTrajectories}
\alias{createTrajectories}
\alias{createTrajectories2}
\title{Spatial Trajectories}
\usage{
createTrajectories(object)

createTrajectories2(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
An updated version of the spata-object specified as \code{object}
now containing the information about all drawn trajectories.
}
\description{
The function \code{createTrajectories()} provides access to an
interactive mini-shiny application that allows to draw trajectories.
}
