% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_documentation.R
\docType{class}
\name{data_counts-class}
\alias{data_counts-class}
\alias{data_counts}
\title{data_counts object}
\value{
S4 object
}
\description{
data_counts object
}
\section{Slots}{

\describe{
\item{\code{counts}}{matrix.}

\item{\code{norm_exp}}{matrix.}
}}

