% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\name{default_parameters}
\alias{default_parameters}
\title{Title}
\usage{
default_parameters()
}
\arguments{
\item{object}{A valid spata-object.}

\item{coords_df}{A data.frame that contains variables \emph{barcodes, sample}.}

\item{features}{The features you want to join \code{coords_df} with specified
as a character vector.}

\item{genes}{The genes you want to join \code{coords_df} with specified
as a character vector.}

\item{average_genes}{Logical value. If set to TRUE coords_df will be joined with
the mean expression values of all genes specified.}

\item{gene_sets}{The gene sets you want to join \code{coords_df} with specified
as a character vector.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{smooth}{Logical value. If set to TRUE values will be
smoothed with respect to their local neighbors using \code{stats::loess()}.}

\item{smooth_span}{Numeric value, given to \code{stats::loess()} if
\code{smooth} is set to TRUE.}

\item{verbose}{Logical value. If set to TRUE informative messages with respect
to the computational progress made will be printed.}
}
\value{

}
\description{
Title
}
\examples{

}
