% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{dim_red-class}
\alias{dim_red-class}
\alias{dim_red}
\title{dim_red object}
\value{
S4 object
}
\description{
dim_red object
}
\section{Slots}{

\describe{
\item{\code{UMAP}}{data.frame.}

\item{\code{TSNE}}{data.frame.}
}}

