% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{discardGeneSets}
\alias{discardGeneSets}
\title{Discard gene sets}
\usage{
discardGeneSets(object, gs_names)
}
\arguments{
\item{object}{A valid spata-object.}

\item{gs_names}{The gene sets to be discarded specified as a character vector.}
}
\value{
An updated spata-object.
}
\description{
Discard gene sets
}
