% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{evaluate_gene_cluster_tendency}
\alias{evaluate_gene_cluster_tendency}
\alias{evaluate_gene_cluster_tendency_dbscan}
\alias{evaluate_gene_cluster_tendency_dbscan2}
\alias{evaluate_gene_cluster_tendency_hdbscan}
\title{Title}
\usage{
evaluate_gene_cluster_tendency(
  marked_df,
  verbose = TRUE,
  pb = NULL,
  ref_totwss,
  ...
)

evaluate_gene_cluster_tendency_dbscan(marked_df, verbose = TRUE, pb = NULL)

evaluate_gene_cluster_tendency_dbscan2(marked_df, verbose = TRUE, pb = NULL)

evaluate_gene_cluster_tendency_hdbscan(marked_df, verbose = TRUE, pb = NULL)
}
\arguments{
\item{pb}{}
}
\value{

}
\description{
Title
}
