% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-expression-analysis.R
\name{filterDE}
\alias{filterDE}
\title{Filter differential genes}
\usage{
filterDE(
  data,
  max_FC = 100,
  min_pvalue = 100,
  across_subset = NULL,
  genes_only = FALSE
)
}
\arguments{
\item{data}{A data.frame containing the numeric variables \emph{p_val, avg_logFC, p_val_adj} and
the character variables \emph{cluster, gene}.}

\item{max_FC}{Numeric value. Denotes the number of genes to keep per \emph{cluster}-group. See details.}

\item{min_pvalue}{Numeric value. Affects the number of genes kept per \emph{cluster}-group. See details.}

\item{across_subset}{Character vector or NULL. Specify the particular groups or clusters of interest the feature-variable
specified in argument \code{across} contains. If set to NULL all of them are chosen.

Hint: Use \code{getFeatureVariables(..., return = "vector", unique = TRUE)} to obtain all available groups of a certain
feature-variable.}

\item{genes_only}{Logical. If set to TRUE the resulting gene-variable is returned as a character vector.}
}
\value{
A filtered data.frame or a character vector of gene names.
}
\description{
Post-process the results of \code{findDE()} according to your needs.
}
