% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-analysis.R
\name{filterTrends}
\alias{filterTrends}
\title{Trajectory trends}
\usage{
filterTrends(atdf, limit = 2, trends = "all")
}
\arguments{
\item{atdf}{An assessed trajectory data.frame (easily accessed via
\code{assessTrajectoryTrends()}).}

\item{limit}{Numeric value. The maximum area-under-the-curve value the
trajectory-trend-assessment might have.}

\item{trend}{Character vector. The patterns of interest.}
}
\value{
A character vector of gene or gene-set names that follow the specified
patterns to the specified degree.
}
\description{
Extracts the genes or gene sets that follow a desired trend.
}
