% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{find_elbow}
\alias{find_elbow}
\title{Cut scree plot at optimal k value}
\usage{
find_elbow(variance)
}
\arguments{
\item{variance}{The sorted variances (totoal within sum of squares).}
}
\description{
Cut scree plot at optimal k value
}
