% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-expression-analysis.R
\name{findDE}
\alias{findDE}
\title{Find marker genes}
\usage{
findDE(
  object,
  of_sample = "",
  across,
  across_subset = NULL,
  method_de = "wilcox",
  p_val_adj = 0.05
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{across}{Character value. The name of the discrete feature-variable of interest.

Hint: Character- or factor- variables are discrete variables. These functionally assign the barcode spots to distinct
groups or clusters (e.g. \emph{segment} or \emph{seurat_clusters}) whoose properties you might want
to compare against each other. Use \code{getFeatureNames()} to get an overview of the
features variables your spata-object contains.}

\item{across_subset}{Character vector or NULL. Specify the particular groups or clusters of interest the feature-variable
specified in argument \code{across} contains. If set to NULL all of them are chosen.

Hint: Use \code{getFeatureValues()} to obtain all available groups of a certain
feature-variable.}

\item{method_de}{Character value. Given to argument \code{test.use} of \code{Seurat::FindAllMarkers()}.}

\item{p_val_adj}{The maximum adjusted p-value allowed in the output.}
}
\value{
A data.frame containing the variables \emph{'p_val', 'avg_logFC', 'pct.1', 'pct.2', 'p_val_adj', 'cluster', 'gene'}.
}
\description{
Finds the differentially expressed genes across a set of subgroups.
}
