% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-expression-analysis.R
\name{findDeGenes}
\alias{findDeGenes}
\title{Find differently expressed genes}
\usage{
findDeGenes(
  object,
  of_sample = "",
  across,
  method_de = "wilcox",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{across}{Character value. The name of the discrete feature-variable of interest.

Hint: Character- or factor- variables are discrete variables. These functionally assign the barcode spots to distinct
groups or clusters (e.g. \emph{segment} or \emph{seurat_clusters}) whoose properties you might want
to compare against each other. Use \code{getFeatureNames()} to get an overview of the
features variables your spata-object contains.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{...}{Additional arguments given to \code{Seurat::FindAllMarkers()}}
}
\value{
A spata-object containing the results in slot @dea.
}
\description{
This function makes use of \code{Seurat::FindAllMarkers()} to compute
the differently expressed genes across the groups denoted in the argument \code{across}.
See details for more.
}
\details{
If \code{across} and/or \code{method_de} are vectors instead of single
values \code{findeDeGenes()} iterates over all combinations in a for-loop and
stores the results in the respective slots. (e.g.: If \code{across} = \emph{'seurat_clusters'}
and \code{method_de} = \emph{c('wilcox', 'bimod')} the function computes the differently expressed genes
across all groups found in the feature variable \emph{seurat_clusters} according to method \emph{wilcox} and
stores the results in the respective slot. Then it does the same according to method \emph{bimod}.)

The results are obtainable via \code{getDeResults()} and \code{getDeGenes()}.
}
