% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern-analysis.R
\name{findIgraphCluster}
\alias{findIgraphCluster}
\title{Clustering with igraph}
\usage{
findIgraphCluster(cor_mtr, num_nn = 20)
}
\arguments{
\item{cor_mtr}{A correlation matrix.}

\item{num_nn}{Numeric value. The maximum number of nearest neighbours to compute.
The default value is set to the smaller of the number of columns in data. Given
to \code{RANN::nn2()} as input for argument \code{k}.}

\item{color_nn}{Numeric value.}

\item{k}{Numeric value.}
}
\value{
Cluster results.
}
\description{
Clustering with igraph
}
