% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-additional.R
\name{findNPatterns}
\alias{findNPatterns}
\title{Title}
\usage{
findNPatterns(object, n_patterns, plotPrSummary = TRUE, of_sample = NA)
}
\arguments{
\item{plotPrSummary}{}
}
\value{

}
\description{
Title
}
