% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-overlap.R
\name{findOverlap}
\alias{findOverlap}
\title{Spatial overlap}
\usage{
findOverlap(
  object,
  of_sample = "",
  input,
  method_gs = "zscore",
  method_ovl = "bayesian",
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  pt_size = 2,
  pt_alpha = 0.9,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{input}{Character vector of length 2. Specifies the genes or gene-sets
whose overlap you want to calculate.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{method_ovl}{Character value. One of \emph{'classic', 'bayesian'}. Decides
according to which method the spatial overlap is calculated.}

\item{pt_clrp}{The color panel to be used if the specified variable displayed by
color is categorical/discrete. Run \code{all_colorpanels()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{all_colorspectra()} to see valid input..}

\item{pt_size}{The size of the points specified as a single numeric value.}

\item{pt_alpha}{The transparency of the points specified as single numeric value.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A named list containing the results (plots, estimated correlation value, data.frames).
}
\description{
-- --
}
