% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform2-family.R
\name{From_Seurat_to_SPATA}
\alias{From_Seurat_to_SPATA}
\title{From_Seurat_to_SPATA}
\usage{
From_Seurat_to_SPATA(seurat_object, sample_names = NULL, verbose = T)
}
\arguments{
\item{sample_names}{if NULL sample name will be created}

\item{verbose}{}
}
\value{
A spata-object.
}
\description{
This Function will return a data frame with the sourounding spots
}
