% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{generate_reference_df}
\alias{generate_reference_df}
\title{Title}
\usage{
generate_reference_df(n, x_coords, y_coords)
}
\arguments{
\item{dropped_df}{dropped df has variables c("x", "y", value)
and is the result of tidyr::drop_na()}
}
\value{

}
\description{
Title
}
