% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getActiveMatrixName}
\alias{getActiveMatrixName}
\title{Obtain name of currently active expression matrix}
\usage{
getActiveMatrixName(object, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Character value.
}
\description{
Obtain name of currently active expression matrix
}
