% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getAvailableDeResults}
\alias{getAvailableDeResults}
\title{Obtain info on de-analysis storage}
\usage{
getAvailableDeResults(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A summarizing list.
}
\description{
Obtain info on de-analysis storage
}
