% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{getBarcodes}
\alias{getBarcodes}
\title{Extract barcodes of a sample}
\usage{
getBarcodes(object, of_sample = "all")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the barcodes.}
}
\value{
All barcodes of the specified sample(s).
}
\description{
Extract barcodes of a sample
}
