% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{getC_joinW_combo}
\alias{getC_joinW_combo}
\title{getC_joinW_combo}
\usage{
getC_joinW_combo()
}
\description{
getC_joinW_combo
}
\seealso{
Combine \code{getCoordinates()} and \code{joinWithGeneSets()} to obtain
a valid input data.frame for \code{data}.
}
