% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getCoordinates}
\alias{getCoordinates}
\alias{getCoordinatesSegment}
\title{Obtain spatial coordinates}
\usage{
getCoordinates(object, of_sample)

getCoordinatesSegment(object, of_segment, of_sample)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}

\item{of_segment}{Character value. Specifies the segment of interest.}
}
\value{
A data.frame containing the variables \emph{barcods, sample, x, y}
(and \emph{segment} if specified).
}
\description{
Obtain spatial coordinates
}
