% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getDeResultDf}
\alias{getDeResultDf}
\title{Obtain de-analysis results}
\usage{
getDeResultDf(
  object,
  of_sample = "",
  across,
  across_subset = NULL,
  method_de = "wilcox",
  max_adj_pval = NULL,
  n_highest_lfc = NULL,
  n_lowest_pval = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{across}{Character value. The name of the discrete feature-variable of interest.

Hint: Character- or factor- variables are discrete variables. These functionally assign the barcode spots to distinct
groups or clusters (e.g. \emph{segment} or \emph{seurat_clusters}) whoose properties you might want
to compare against each other. Use \code{getFeatureNames()} to get an overview of the
features variables your spata-object contains.}

\item{across_subset}{Character vector or NULL. Specify the particular groups or clusters of interest the feature-variable
specified in argument \code{across} contains. If set to NULL all of them are chosen.

Hint: Use \code{getFeatureValues()} to obtain all available groups of a certain
feature-variable.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{max_adj_pval}{Numeric value. Sets the maximal threshold for adjusted p-values allowed. All genes
with adjusted p-values above that threshold are ignored.}

\item{n_highest_lfc}{Numeric value. Affects the total number of genes that are kept. See details.}

\item{n_lowest_pval}{Numeric value. Affects the total number of genes that are kept. See details.}
}
\value{
A data.frame:

\itemize{
  \item{\emph{gene}} Character. The differentially expressed genes.
  \item{\emph{cluster}} Character. The clusters (or experimental groups) across which the analysis was performed.
  \item{\emph{avg_logFC}} Numeric. The average log-fold change to which the belonging gene was differentially expressed..
  \item{\emph{p_val}} Numeric. The p-values.
  \item{\emph{p_val_adj}} Numeric. The adjusted p-values.
 }

If \code{getDeGenes()} is used the \emph{gene}-variable is returned as a named character vector.
}
\description{
Obtain de-analysis results
}
\details{
The de-data.frame is processed such that the following steps are performed for every experimental
group. (With "genes" we refer to the rows (observations) of \code{data}.)

\enumerate{
 \item{Discards genes with \emph{avg_logFC}-values that are either infinite or negative}
 \item{Discards genes with adjusted p-values above the threshold set with \code{max_adj_pval}}
 \item{Slices the data.frame in order that for every unique cluster of the \emph{cluster}-variable}:
 \enumerate{
  \item{the n genes with the highest \emph{avg_logFC}-values are kept where n = \code{n_highest_lfc}}
  \item{the n genes with the lowest \emph{p_val_adj}-values are kept where n = \code{n_lowest_pval}}
  }
 \item{Arranges the genes according to the highest \emph{avg_logFC}-values}
 }
}
