% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getDeResults}
\alias{getDeResults}
\alias{getDeGenes}
\title{Obtain de-analysis results}
\usage{
getDeResults(
  object,
  of_sample = "",
  across,
  method_de,
  p_val_adj = 0.5,
  n_threshold = c(50, 50)
)

getDeGenes(
  object,
  of_sample = "",
  across,
  method_de = "wilcox",
  p_val_adj = NULL,
  n_threshold = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{across}{Character value. The name of the discrete feature-variable of interest.

Hint: Character- or factor- variables are discrete variables. These functionally assign the barcode spots to distinct
groups or clusters (e.g. \emph{segment} or \emph{seurat_clusters}) whoose properties you might want
to compare against each other. Use \code{getFeatureNames()} to get an overview of the
features variables your spata-object contains.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{p_val_adj}{Numeric value. Denotes the max. adjusted p-value a gene can have
in order to be kept in the results.}

\item{n_threshold}{Numeric value or vector of length two. See details for more.}
}
\value{
A data.frame:

\itemize{
  \item{\emph{gene}} Character. The differentially expressed genes.
  \item{\emph{cluster}} Character. The clusters (or experimental groups) across which the analysis was performed.
  \item{\emph{avg_logFC}} Numeric. The average log-fold change to which the belonging gene was differentially expressed..
  \item{\emph{p_val}} Numeric. The p-values.
  \item{\emph{p_val_adj}} Numeric. The adjusted p-values.
 }

If \code{getDeGenes()} is used the \emph{gene}-variable is returned as a named character vector.
}
\description{
Obtain de-analysis results
}
\details{
De-analysis results are stored in a data.frame in which each gene that turned out
to be a marker gene for a specific group is evaluated by it's p-value, p-adjusted-value,
logfold-change etc.

If the arguments \code{p_val_adj} and \code{n_threshold} are set to NULL the results are returned
as they are which however can result in several hundred genes per group. The results can be additionally
subsetted. \code{getDeResults()} does that by

\enumerate{
 \item{discarding genes with \emph{avg_logFC}-values that are either infinite or negative}
 \item{discarding genes with a p-adjusted-value higher than the threshold denoted in \code{p_val_adj}}
 \item{slicing the data.frame in order that for every unique group of the feature denoted in \code{across}}:
 \enumerate{
  \item{the n genes with the highest \emph{avg_logFC}-values are kept where n = \code{n_threshold[1]}}
  \item{the n genes with the lowest \emph{p_val_adj}-values are kept where n = \code{n_threshold[2]}}
  }
 \item{arranging the genes according to the highest \emph{avg_logFC}-values}
 }
}
