% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getDimRedData}
\alias{getDimRedData}
\alias{getUmapData}
\alias{getTsneData}
\title{Obtain dimensional reduction data}
\usage{
getDimRedData(object, of_sample, method_dr = c("UMAP", "TSNE"))

getUmapData(object, of_sample)

getTsneData(object, of_sample)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}

\item{method_dr}{The dimensional reduction method of
interest specified as a single character value. (Currently
either \emph{'UMAP'} or \emph{'TSNE'}).}
}
\value{
A data.frame that contains the unique identifiers
(keys): \emph{barcodes, sample} and:.

 \itemize{
  \item{ \code{getTsneData()}: \emph{tsne1, tsne2}}
  \item{ \code{getUmapData()}: \emph{umap1, umap2}}
  }
}
\description{
Obtain dimensional reduction data
}
