% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\title{Obtain expression matrix}
\usage{
getExpressionMatrix(object, of_sample = "all")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) from which to extract the expression data
specified as a character vector.}
}
\value{
The expression matrix of the specified object.
}
\description{
Obtain expression matrix
}
