% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFeatureData}
\alias{getFeatureData}
\title{Obtain feature data}
\usage{
getFeatureData(object, of_sample)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}
}
\value{
The feature data data.frame of the specfied object and sample(s).
}
\description{
Obtain feature data
}
