% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFeatureNames}
\alias{getFeatureNames}
\title{Obtain feature names}
\usage{
getFeatureNames(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A named character vector of the variables in the feature data slot.
}
\description{
Obtain feature names
}
