% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFeatureVariables}
\alias{getFeatureVariables}
\title{Obtain a feature variable}
\usage{
getFeatureVariables(
  object,
  features,
  of_sample = "",
  return = "data.frame",
  unique = FALSE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{features}{The features of interest specified as a character vector.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{return}{Character value. Determine the way the variable is returned.

(Only relevant if input of arugment \code{features} is of length greater than 1.)}

\item{unique}{Logical. If set to TRUE and argument \code{features} is
of length 1 only it's unique values are returned.}
}
\value{
A data.frame or a vector.
}
\description{
Extracts the specified feature variables from the
feature data.
}
