% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGeneHotspotDf}
\alias{getGeneHotspotDf}
\title{Obtain gene hotspot evaluation data frame}
\usage{
getGeneHotspotDf(object, of_sample = "", genes = NULL, mtr_name = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{genes}{NULL or character vector. Denotes the genes of interest. The data.frame
is filtered for them in addition to beeing filtered for those gene that
were included in hotspot analysis process. Ignored if set to NULL.}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}
}
\value{
A nested data.frame.
}
\description{
Obtain gene hotspot evaluation data frame
}
