% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGeneMetaData}
\alias{getGeneMetaData}
\alias{getGeneMetaDf}
\title{Obtain gene meta data}
\usage{
getGeneMetaData(object, of_sample = "", mtr_name = NULL, only_df = FALSE)

getGeneMetaDf(object, of_sample = "", mtr_name = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}

\item{only_df}{Logical. If set to TRUE only the data.frame is returned.
If set to FALSE (the default) the whole list is returned.}
}
\value{
A data.frame from \code{getMetaDataDf()} or a list from \code{getGeneMetaData()}.
}
\description{
Obtain gene meta data
}
