% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getHotspotInfoDf}
\alias{getHotspotInfoDf}
\alias{getHotspotDf}
\title{Obtain pattern recognition results of method 'hotspot'}
\usage{
getHotspotInfoDf(object, of_sample = NA)

getHotspotDf(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
The respective data.frame.
}
\description{
Obtain pattern recognition results of method 'hotspot'
}
