% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRNAvelocity.R
\name{getRNAvelocity}
\alias{getRNAvelocity}
\title{getRNAvelocity}
\usage{
getRNAvelocity(
  object,
  Velocity_from = c("spatial", "UMAP"),
 
    path_to_pythonscript = "'/Users/HenrikHeiland/Desktop/T-Cell Project/Revisions/sc2velo.py'",
  loom,
  path_to_python = "/Users/HenrikHeiland/opt/anaconda3/bin/python3.7 ",
  folder = getwd()
)
}
\arguments{
\item{Velocity_from}{can be "spatial" if you aim to project in the spatial space or a "UMAP")}

\item{path_to_pythonscript}{The py skript sc2velo.py is required and in in the data folder of the SPATA github}

\item{loom}{loom file}

\item{path_to_python}{where is you python 3.8 (with installed scvelo)}

\item{folder}{folder where files will be saved}
}
\value{
A spata-object.
}
\description{
This Function will return a SPATA object containing denoised data and re-analysis
}
