% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSamples}
\alias{getSamples}
\title{Obtain sample names}
\usage{
getSamples(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A character vector.
}
\description{
Obtain sample names
}
