% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSegmentNames}
\alias{getSegmentNames}
\title{Obtain segment names}
\usage{
getSegmentNames(object, of_sample = "all")
}
\arguments{
\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}
}
\value{
A list named according to the \code{of_sample} in which each element is
a character vector containing the names of segments which were drawn for the
specific sample.
}
\description{
Obtain segment names
}
