% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSpCorCluster}
\alias{getSpCorCluster}
\title{Title}
\usage{
getSpCorCluster(object, method_hclust = "complete", of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{method_hclust}{Character value. Denotes the method that was used to generate the
clustering results you want to extract.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{

}
\description{
Title
}
