% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSpCorResults}
\alias{getSpCorResults}
\title{Title}
\usage{
getSpCorResults(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{

}
\description{
Title
}
