% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSummarizedTrajectoryDf}
\alias{getSummarizedTrajectoryDf}
\title{Obtain a summarized trajectory data.frame}
\usage{
getSummarizedTrajectoryDf(object, trajectory_name, of_sample, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}

\item{...}{Additional arguments given to \code{hlpr_summarize_trajectory_df()}.}
}
\value{
A summarized trajectory data.frame.
}
\description{
A wrapper arround \code{getTrajectoryObject()} and
\code{hlpr_sumamrize_trajectory_df()}.
}
