% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family_new.R
\name{getSurroundedSpots}
\alias{getSurroundedSpots}
\title{getSurroundedSpots}
\usage{
getSurroundedSpots(object, of_sample = NULL, assay_type = "Visium")
}
\arguments{
\item{of_sample}{the sample to be analyzed}

\item{assay_type}{Visium with 6 sourounding spots and ST with 8 spots}
}
\value{
A spata-object.
}
\description{
This Function will return a data frame with the sourounding spots
}
