% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTrajectoryLength}
\alias{getTrajectoryLength}
\title{Obtain the length of a trajectory}
\usage{
getTrajectoryLength(object, trajetory_name, of_sample, binwidth)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}
}
\value{
Numeric value.
}
\description{
This function returns the length (the number of bins) of a trajectory
depending on the chosen \code{binwidth}.
}
