% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTrajectoryNames}
\alias{getTrajectoryNames}
\title{Obtain trajectory names}
\usage{
getTrajectoryNames(object, of_sample = "all", simplify = TRUE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}
}
\value{
A list named according to the \code{of_sample} in which each element is
a character vector containing the names of trajectories which were drawn for the
specific sample.
}
\description{
Obtain trajectory names
}
