% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_assign}
\alias{hlpr_assign}
\title{Assign objects into the global environment}
\usage{
hlpr_assign(assign, object, name)
}
\arguments{
\item{assign}{Logical.}

\item{object}{The object to be assigned.}

\item{name}{The name of the assigned object.}
}
\description{
Assign objects into the global environment
}
