% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_compare_samples}
\alias{hlpr_compare_samples}
\title{Compare samples in df with those from object}
\usage{
hlpr_compare_samples(object, df, messages)
}
\arguments{
\item{object}{A valid spata-object.}

\item{df}{A data.frame}

\item{messages}{A message vector.}
}
\value{
Updated message vector
}
\description{
Compare samples in df with those from object
}
