% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_compile_trajectory}
\alias{hlpr_compile_trajectory}
\title{Compiles a trajectory data.frame}
\usage{
hlpr_compile_trajectory(
  segment_trajectory_df,
  trajectory_width,
  object,
  sample
)
}
\arguments{
\item{segment_trajectory_df}{A data.frame specifying each segment of the whole
trajectory with variables \code{x, y, xend, yend}.}

\item{trajectory_width}{Numeric value that determines the width of the
trajectory.}

\item{object}{A valid spata-object.}

\item{sample}{The sample specified as a character string of length one.}
}
\value{
A data.frame containing the variables \emph{barcodes, sample, x, y}
as well as \emph{projection_length} indicating the position of the barcode
with respect to the direction of the trajectory-part and \emph{trajectory_part}
indicating the part of the trajectory the barcode belongs to.
}
\description{
Compiles a trajectory data.frame
}
