% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_compute_eps}
\alias{hlpr_compute_eps}
\title{Title}
\usage{
hlpr_compute_eps(object, of_sample = "")
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Compute optimal eps input for \code{dbsan::dbscan()}.
}
