% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{hlpr_image_add_on}
\alias{hlpr_image_add_on}
\title{}
\usage{
hlpr_image_add_on(object, display_image)
}
\arguments{
\item{object}{A valid spata-object.}

\item{display_image}{Logical value.}
}
\value{
Either null or a ggplot2::geom_annotation_raster
}
\description{

}
