% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_lm_info}
\alias{hlpr_lm_info}
\title{Linear model summary}
\usage{
hlpr_lm_info(model)
}
\arguments{
\item{lm}{A linear model object.}
}
\value{
A data.frame of one row.
}
\description{
Takes a linear model and returns an assembled data.frame
of one row containing information about the models slope, adjusted Rsquared
and pvalue.
}
