% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_lm_trajectory}
\alias{hlpr_lm_trajectory}
\title{Trajectory - Linear model}
\usage{
hlpr_lm_trajectory(df)
}
\arguments{
\item{df}{A trajectory data.frame that was generated by
\code{hlpr_summarize_trajectory_df()}.}
}
\value{
A liner model.
}
\description{
To be used in \code{rankTrajectoryGenes()} or
in \code{rankTrajectoryGenes()} as input for \code{.f}-argument in
\code{purrr::map()}. Models the expression of the respective
gene set expression values along the trajectory order of
the provided data.frame.
}
