% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_smooth}
\alias{hlpr_smooth}
\title{Smooth variables spatially}
\usage{
hlpr_smooth(
  variable,
  var_name,
  coords_df,
  verbose = TRUE,
  smooth_span,
  aspect,
  subset
)
}
\arguments{
\item{variable}{The variable to smooth}

\item{var_name}{Name of the variable to smooth}

\item{coords_df}{Data.frame that contains x and y coordinates}

\item{verbose}{Logical}

\item{smooth_span}{Span to smooth with}

\item{aspect}{Gene or Gene set}

\item{subset}{Vector of variable names to smooth}
}
\value{
Smoothed variable (data.frame within \code{purrr::imap()})
}
\description{
Helper function to use within \code{purrr::imap()}
}
