% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_summarize_trajectory_df}
\alias{hlpr_summarize_trajectory_df}
\title{Summarize and join ctdf}
\usage{
hlpr_summarize_trajectory_df(
  object,
  ctdf,
  accuracy = 5,
  variables,
  method_gs = "mean",
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{ctdf}{A compiled trajectory data.frame containing the variables
\emph{'barcodes', 'sample', 'x', 'y', 'projection_length', 'trajectory_part'}.}

\item{accuracy}{Numeric. Given to \code{accuracy}-argument of
\code{plyr::round_any()}. Determines how many barcode-spots will be summarized
as one sub-trajectory-part.}

\item{variables}{The variables of interest specified as a character vector:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 \item{ \strong{Features}: Must be a numeric one of \code{getFeaturenNames()}}
 }}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Joins a compiled trajectory data.frame with
the desired information and summarizes it's joined values
along the sub-trajectory-parts via \code{base::mean()}.
}
