% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern-analysis.R
\name{igraph_cluster}
\alias{igraph_cluster}
\alias{plot_igraph_heatmap}
\title{Clustering with igraph}
\usage{
igraph_cluster(cor.mtr, num.nn = 20)

plot_igraph_heatmap(cor.mtr, num.nn, color.nn)
}
\arguments{
\item{cor.mtr}{A correlation matrix.}

\item{num.nn}{Numeric value.}

\item{color.nn}{Numeric value.}
}
\value{
Cluster results.
}
\description{
Clustering with igraph
}
