% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generic-functions.R
\name{image,spata-method}
\alias{image,spata-method}
\title{Methods}
\usage{
\S4method{image}{spata}(object, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the content.}

\item{trajectory_name}{The trajectory specified as a character value.}
}
\description{
Methods
}
