% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generic_functions.R
\name{image}
\alias{image}
\alias{exprMtr}
\alias{coordinates}
\alias{coordinates<-}
\alias{featureData}
\alias{featureData<-}
\alias{samples}
\alias{trajectory}
\alias{getTrajectoryComment}
\alias{image,spata-method}
\alias{exprMtr,spata-method}
\alias{coordinates,spata-method}
\alias{coordinates<-,spata-method}
\alias{featureData,spata-method}
\alias{featureData<-,spata-method}
\alias{samples,spata-method}
\alias{trajectory,spata-method}
\alias{getTrajectoryComment,spata-method}
\alias{getTrajectoryComment,spatialTrajectory-method}
\title{Generics to extract a slots content}
\usage{
image(object, of_sample)

exprMtr(object, of_sample = NULL)

coordinates(object, of_sample = "all")

coordinates(object) <- value

featureData(object, of_sample = "all")

featureData(object) <- value

samples(object)

trajectory(object, trajectory_name, of_sample)

getTrajectoryComment(object, ...)

\S4method{image}{spata}(object, of_sample)

\S4method{exprMtr}{spata}(object, of_sample = "all")

\S4method{coordinates}{spata}(object, of_sample = "all")

\S4method{coordinates}{spata}(object) <- value

\S4method{featureData}{spata}(object, of_sample = "all")

\S4method{featureData}{spata}(object) <- value

\S4method{samples}{spata}(object)

\S4method{trajectory}{spata}(object, trajectory_name, of_sample)

\S4method{getTrajectoryComment}{spata}(object, trajectory_name, of_sample)

\S4method{getTrajectoryComment}{spatialTrajectory}(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the content.}

\item{trajectory_name}{The trajectory specified as a character value.}
}
\value{
The respective slots content.
}
\description{
Generics to extract a slots content
}
