% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate-family.R
\name{initiateSpataObject_ExprMtr}
\alias{initiateSpataObject_ExprMtr}
\title{Initiate spata object from scaled expression matrix}
\usage{
initiateSpataObject_ExprMtr(
  coords_df,
  expr_mtr,
  count_mtr = NULL,
  mtr_name = "scaled",
  sample_name = "sample1",
  image = NULL,
  gene_set_path = NULL,
  output_path = NULL,
  file_name = NULL,
  pca_comp = 30,
  k = 50,
  tsne_perplexity = 30,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{coords_df}{Data.frame containing information about the positions of all
barcode-spots in form of a numeric \emph{x}- and \emph{y}-variable. The key-variable
\emph{barcodes} needs to be of type character and must be identical to the column names
of the input matrix (\code{expr_mtr}).}

\item{expr_mtr}{A numeric matrix. The expression matrix to be used.}

\item{sample_name}{Character value. The future input for SPATA's \code{of_sample}-argument.}

\item{image}{An Image of the sample that can be displayed as the surface plot's background.}

\item{gene_set_path}{Character value (or NULL). Specifies the path to a
.RDS-file containing a data.frame that is to be used as input for slot @used_genesets.

 Must have the character-variables

   \itemize{
    \item{\emph{'ont'}: The gene set names.}
    \item{\emph{'gene'}: The belonging gene names.}
    }

If set to NULL the default gene-set data.frame will used. Run \code{?gsdf} to get more information.}

\item{output_path}{Character vector or NULL. Specifies the folder in which to store
the object if the directory is valid. If set to NULL saving is skipped.}

\item{file_name}{Character value. The name-suffix for the file name under which the
spata-object is stored if \code{output_path} is a valid directory. Is prefixed with
 \emph{'spata-obj-'} and suffixed with \emph{'.RDS'}.}

\item{pca_comp}{Numeric value. Given to argument \code{n} of function \code{irlba::prcomp_irlba()}: Determines
the number of components to return.}

\item{tsne_perplexity}{Numeric value. Given to argument \code{perplexity} of function \code{Rtsne::Rtsne()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{nn}{Numeric value. Given to argument \code{k} of function \code{RANN::nn2()}: Determines to maximum number
of nearest neighbours to compute.}
}
\value{
A spata-object.
}
\description{
Default function for any spatial related experiment whoose output is
an already processed expression/intensity matrix. See details for more information.
}
\details{
After initiating the spata-object PCA is performed via \code{irlba::prcomp_irlba()} and clustering
is done via \code{RANN::nn2()}. (Use \code{addFeatures()} to add any clustering results of your own analysis.)
Additional dimensional reduction is performed via \code{Rtsne::Rtsne()} and \code{umap::umap()}.
}
