% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate-family.R
\name{initiateSpataObject_MALDI}
\alias{initiateSpataObject_MALDI}
\title{Initiate a spata-object from MALDI Experiments}
\usage{
initiateSpataObject_MALDI(
  coords_df,
  intensity_mtr,
  sample_name,
  gene_set_path = NULL,
  output_path = NULL,
  file_name = NULL,
  pca_comp = 30,
  nn = 50,
  tsne_perplexity = 30,
  verbose = TRUE
)
}
\arguments{
\item{intensity_mtr}{A numeric matrix to be used as the expression matrix. Rownames must
correspond to the genes and column names must correspond to the barcodes.}
}
\value{
A spata-object.
}
\description{
Initiate a spata-object from MALDI Experiments
}
