% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{iterate_over_kmeans}
\alias{iterate_over_kmeans}
\title{Title}
\usage{
iterate_over_kmeans(
  dropped_df,
  vars = c("x", "y"),
  max_cluster = 9,
  n_start = 10,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
