% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinWith-family.R
\name{joinWith}
\alias{joinWith}
\alias{joinWithFeatures}
\alias{joinWithGenes}
\alias{joinWithGeneSets}
\alias{joinWithVariables}
\title{Join barcodes with additional variables}
\usage{
joinWith(
  object,
  coords_df,
  features = NULL,
  method_gs = "mean",
  gene_sets = NULL,
  genes = NULL,
  average_genes = FALSE,
  smooth = FALSE,
  smooth_span = 0.02,
  verbose = TRUE,
  normalize = TRUE
)

joinWithFeatures(
  object,
  coords_df,
  features,
  smooth = FALSE,
  smooth_span = 0.02,
  verbose = TRUE
)

joinWithGenes(
  object,
  coords_df,
  genes,
  average_genes = FALSE,
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE
)

joinWithGeneSets(
  object,
  coords_df,
  gene_sets,
  method_gs = "mean",
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE
)

joinWithVariables(
  object,
  coords_df,
  variables,
  method_gs = "mean",
  average_genes = FALSE,
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{coords_df}{A data.frame that contains at least the character variables
\emph{samples} and \emph{barcodes}.}

\item{features}{The features of interest specified as a character vector.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{gene_sets}{The gene sets of interest specified as a character vector.}

\item{genes}{The genes of interest specified as a character vector.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{variables}{Character vector. The variables of interest:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 \item{ \strong{Features}: Must be numeric ones of \code{getFeatureNames()}}
 }}
}
\value{
The input data.frame of \code{coords_df} joined with all the
specified variable-elements (by the key-variable \emph{barcodes}).
}
\description{
Each member of the \code{joinWith()}-family takes a data.frame as
input that contains at least the variables \emph{barcodes} and \emph{sample}.
(Easily obtained with the \code{get*()}-family.) It extracts the specified
variables and joins them over the barcode variable with the provided data.frame.
}
