% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinWith_family.R
\name{joinWithFeatures}
\alias{joinWithFeatures}
\alias{joinWithGenes}
\alias{joinWithGeneSets}
\title{Join barcodes with additional variables}
\usage{
joinWithFeatures(
  object,
  coords_df,
  features,
  smooth = FALSE,
  smooth_span = 0.02,
  verbose = TRUE
)

joinWithGenes(
  object,
  coords_df,
  genes,
  average_genes = FALSE,
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE
)

joinWithGeneSets(
  object,
  coords_df,
  gene_sets,
  method_gs = "mean",
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{coords_df}{A data.frame that contains variables \emph{barcodes, sample}}

\item{features}{The features you want to join \code{coords_df} with specified
as a character vector.}

\item{smooth}{Logical value. If set to TRUE values will be
smoothed with respect to their local neighbors using \code{stats::loess()}.}

\item{smooth_span}{Numeric value, given to \code{stats::loess()} if
\code{smooth} is set to TRUE.}

\item{verbose}{Logical value. If set to TRUE informative messages with respect
to the computational progress made will be printed.

(Warning messages will always be printed.)}

\item{genes}{The genes you want to join \code{coords_df} with specified
as a character vector.}

\item{average_genes}{Logical value. If set to TRUE coords_df will be joined with
the mean expression values of all genes specified.}

\item{gene_sets}{The gene sets you want to join \code{coords_df} with specified
as a character vector.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}
}
\value{
The input data.frame of \code{coords_df} joined with all the aspects
specified variables specified in \code{features} by the key \emph{barcodes}.
}
\description{
Each member of the \code{joinWith()}-family takes a data.frame as
input that contains at least the variables \emph{barcodes} and \emph{sample}.
(Easily obtained with the \code{spata::coords()}-family.) It joins them over
 the \emph{barcodes}-variable with a specified aspect.
}
