% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{loadGSDF}
\alias{loadGSDF}
\title{Original load gene set data.frame}
\usage{
loadGSDF(gene_set_path = NULL, verbose = TRUE)
}
\description{
Not exported due to naming issues. Kept as it is used in several
loading functions.
}
