% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{loadSpataObject}
\alias{loadSpataObject}
\title{Load a spata-object}
\usage{
loadSpataObject(path)
}
\arguments{
\item{path}{Character value. The file directory.}
}
\value{
The loaded object.
}
\description{
A wrapper around \code{base::readRDS()}.
}
