% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{mark_with_na}
\alias{mark_with_na}
\alias{mark_all_with_na}
\alias{mark_with_na2}
\alias{mark_all_with_na2}
\title{Mark values of variables with NA}
\usage{
mark_with_na(variable, n_qntls = 5, keep_qntls = 5)

mark_all_with_na(coords_df, n_qntls = 5, keep_qntls = 5)

mark_with_na2(variable, percentile = 0.75)

mark_all_with_na2(coords_df, percentile = 0.75)
}
\arguments{
\item{variable}{The numeric variable.}

\item{n_qntls}{Numeric value. Denotes the number of quantiles
the variable's values are binned into.}

\item{keep_qntls}{Numeric vector. Inicates the quantiles
whose values are kept - not marked with NA.}
}
\description{
Mark values of variables with NA
}
