% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_surface_plot.R
\name{moduleSurfacePlotUI}
\alias{moduleSurfacePlotUI}
\title{UI of surface plot module}
\usage{
moduleSurfacePlotUI(id)
}
\arguments{
\item{id}{The namespace id.}
}
\description{
UI of surface plot module
}
