% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{plot_family}
\alias{plot_family}
\title{plot_family}
\usage{
plot_family()
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
plot_family
}
