% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotAutoencoderResults}
\alias{plotAutoencoderResults}
\title{Plot scaled vs. denoised expression}
\usage{
plotAutoencoderResults(
  object,
  genes,
  mtr_name = "denoised",
  scales = "free",
  pt_size = NULL,
  pt_alpha = NULL,
  pt_clrp = NULL,
  verbose = NULL,
  of_saple = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{genes}{Character vector of length two. Denotes the genes to be used for the validation plot.}

\item{pt_size}{Numeric value. Specifies the size of points.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}
}
\description{
Compares the distribution of the expression levels of \code{genes}
between the scaled matrix and the denoised matrix.
}
\details{
This function requires a denoised matrix in slot @data generated
by \code{runAutoEncoderDenoising()} as well as a scaled matrix.
}
