% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-summaries.R
\name{plotDeaGeneCount}
\alias{plotDeaGeneCount}
\title{Title}
\usage{
plotDeaGeneCount(
  object,
  across = NULL,
  across_subset = NULL,
  relevel = FALSE,
  method_de = NULL,
  max_adj_pval = NULL,
  clrp = NULL,
  display_title = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
