% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-summaries.R
\name{plotDeaPvalues}
\alias{plotDeaPvalues}
\alias{plotDeaLogFC}
\title{Title}
\usage{
plotDeaPvalues(
  object,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  method_de = NULL,
  max_adj_pval = NULL,
  binwidth = NULL,
  clrp = NULL,
  plot_type = "histogram",
  scales = NULL,
  limits_x = c(NA, NA),
  display_title = NULL,
  of_sample = NA,
  ...
)

plotDeaLogFC(
  object,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  method_de = NULL,
  max_adj_pval = NULL,
  binwidth = NULL,
  clrp = NULL,
  plot_type = "histogram",
  scales = NULL,
  limits_x = c(NA, NA),
  display_title = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
