% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-summaries.R
\name{plotDeaSummary}
\alias{plotDeaSummary}
\title{Title}
\usage{
plotDeaSummary(
  object,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  method_de = NULL,
  max_adj_pval = NULL,
  clrp = NULL,
  plotDeaGeneCount = list(display_title = TRUE),
  plotDeaLogFC = list(display_title = FALSE),
  plotDeaPvalues = list(display_title = FALSE),
  verbose = NULL,
  of_sample = NA
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
