% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDistributionAcross}
\alias{plotDistributionAcross}
\alias{plotDistributionAcross2}
\title{Distribution of continuous values}
\usage{
plotDistributionAcross(object, of_sample = "", variables, across,
  across_subset = NULL, method_gs = "mean", plot_type = "violin",
  binwidth = 0.05, clrp = "milo", ..., normalize = TRUE,
  assign = FALSE, assign_name, verbose = TRUE)

plotDistributionAcross2(df, variables = "all", across,
  across_subset = NULL, plot_type = "violin", binwidth = 0.05,
  clrp = "milo", ..., normalize = TRUE, assign = FALSE, assign_name,
  verbose = TRUE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{variables}{Character vector. The numeric variables of interest.}

\item{across}{Character value. The name of the discrete feature-variable of interest.

Hint: Character- or factor- variables are discrete variables. These functionally assign the barcode spots to distinct
groups or clusters (e.g. \emph{segment} or \emph{seurat_clusters}) whoose properties you might want
to compare against each other. Use \code{getFeatureNames()} to get an overview of the
features variables your spata-object contains.}

\item{across_subset}{Character vector or NULL. Specify the particular groups or clusters of interest the feature-variable
specified in argument \code{across} contains. If set to NULL all of them are chosen.

Hint: Use \code{getFeatureValues()} to obtain all available groups of a certain
feature-variable.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{plot_type}{Character value. One of \emph{'histogram', 'density', 'violin', 'boxplot' and 'ridgeplot'}.}

\item{binwidth}{The binwidth to use if \code{plot_type} is specified as \emph{'histogram'}.}

\item{...}{additional arguments to \code{ggplot2::facet_wrap()}}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{assign}{Logical. If set to TRUE a named list will be assigned to the global
environment. This list contains data and information to rebuild or additionally
customize the output plot of this function.}

\item{assign_name}{The name the assigned list is supposed to have specified as
a single character value.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{df}{A data.frame that contains the numeric variables specified in \code{variables}.}
}
\description{
Visualizes the distribution of values of a set of variables for the
whole sample across specific subgroups.

\itemize{
 \item{ \code{plotDistributionAcross()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters.}
 \item{ \code{plotDistributionAcross2()} Takes a data.frame as the starting point.}
 }
}
