% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDistributionDiscrete}
\alias{plotDistributionDiscrete}
\title{Distribution of discrete features}
\usage{
plotDistributionDiscrete(
  object,
  of_sample = "",
  discrete_x,
  discrete_y,
  clrp = "milo",
  position = "fill",
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{discrete_x}{Character value. The discrete feature of interest to be mapped onto the x-axis.}

\item{discrete_y}{Character value. The discrete feature of interest to be mapped onto the y-axis.}

\item{clrp}{clrp params}

\item{position}{Character value. Given to \code{position} of \code{ggplot2::geom_bar()}.}

\item{...}{Additional parameter given to \code{ggplot2::geom_bar()}.}
}
\description{
Visualize the distribution of two grouping variables with
a barplot.
}
