% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotExpression}
\alias{plotExpression}
\title{Visualize variable distribution}
\usage{
plotExpression(
  data,
  plot_type = "histogram",
  variables = NULL,
  binwidth = 0.05
)
}
\arguments{
\item{data}{The data.frame containing numeric variables.}

\item{plot_type}{One of \emph{'histogram', 'density', 'ridgeplot', 'boxplot', 'violin'.}}

\item{variables}{The numeric variable whose distribution you want to display.}

\item{binwidth}{The binwidth to use if \code{plot_type} is  \emph{'histogram'}
specified as a numeric value.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Visualize variable distribution
}
\details{
Convenient to use with the data.frames resulting from \code{coordsSpatial() + joinWith*()}.
}
