% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotFourStates}
\alias{plotFourStates}
\alias{plotFourStates2}
\title{Gene set state plot}
\usage{
plotFourStates(
  data,
  states,
  color_to = NULL,
  pt_size = 1.5,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  xlim = c(-1, 1),
  ylim = c(-1, 1)
)

plotFourStates2(
  object,
  of_sample,
  states,
  color_to = NULL,
  method_gs = "gsva",
  pt_size = 1.5,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  verbose = TRUE
)
}
\arguments{
\item{data}{A data.frame containing at least the variables \emph{barcodes, \code{states.}}.
Whereby the states-variables contain the respective expression values of the specified
gene sets.}

\item{states}{The gene sets defining the four states specified as a character vector
of length 4.}

\item{color_to}{The variable in the data frame that is supposed to be displayd by color
specified as a character value.}

\item{pt_size}{The size of the points specified as a numeric value.}

\item{pt_alpha}{The transparency of the points specified as a numeric value.}

\item{pt_clrsp}{The colour spectrum used to display \code{color_to} if the
specified variable is continuous. Needs to be one of \emph{inferno, magma,
plasma, cividis or viridis}.}

\item{xlim}{The limits of the x-axis specified as a numeric vector of length 2.}

\item{ylim}{The limits of the y-axis specified as a numeric vector of length 2.}

\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the data specified as a
character value.}

\item{verbose}{Logical value. If set to TRUE informative messages with respect
to the computational progress made will be printed.

(Warning messages will always be printed.)}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
This function takes four gene sets and visualizes the relative
expression of these four gene sets for every barcode by computing it's respective
x- and y- coordinates in the state plot. (See details.)

(\code{plotFourStates2()} generates the data.frame that needs to be specified
in \code{data} from scratch. It then calls \code{plotFourStates()}).
}
