% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotGeneCenter}
\alias{plotGeneCenter}
\title{Title}
\usage{
plotGeneCenter(
  object,
  of_sample = "",
  of_gene,
  pt_clr = "black",
  pt_shape = 19,
  pt_size = 8,
  plotSurface = FALSE
)
}
\arguments{
\item{plotSurface}{}
}
\value{

}
\description{
Title
}
