% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotGeneExpression}
\alias{plotGeneExpression}
\alias{plotGeneSetExpression}
\title{Visualize expression values}
\usage{
plotGeneExpression(
  object,
  of_sample,
  genes,
  plot_type = "histogram",
  binwidth = 0.05,
  normalize = TRUE,
  verbose = TRUE
)

plotGeneSetExpression(
  object,
  of_sample,
  gene_sets,
  method_gs = "mean",
  plot_type = "histogram",
  binwidth = 0.05,
  normalize = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the data specified as a
character value.}

\item{genes}{The genes of interest specified as a character vector.}

\item{plot_type}{One of \emph{'histogram', 'density', 'ridgeplot', 'boxplot', 'violin'.}}

\item{binwidth}{The binwidth to use if \code{plot_type} is  \emph{'histogram'}
specified as a numeric value.}

\item{verbose}{Logical value to specify whether informative messages are
supposed to be printed or not.}

\item{gene_sets}{The gene sets of interest specified as a character vector.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}. Ignored if \code{color_to} isn't a gene set.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Visualizes the distribution of the chosen genes or
gene sets via a variety of plots.
}
