% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotGeneMetaData}
\alias{plotGeneMetaData}
\title{Title}
\usage{
plotGeneMetaData(
  object,
  of_sample = "",
  variables,
  pt_size = 1.5,
  pt_alpha = 0.8,
  pt_clr = "black",
  mtr_name = NULL
)
}
\arguments{
\item{mtr_name}{}
}
\value{

}
\description{
Title
}
