% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotPatternEnrichment}
\alias{plotPatternEnrichment}
\title{Title}
\usage{
plotPatternEnrichment(
  object,
  of_sample = "",
  of_hotspots = "",
  of_gene_sets = "",
  clrp = "milo",
  top_n = 10,
  with_ties = FALSE,
  ...
)
}
\arguments{
\item{with_ties}{}
}
\value{

}
\description{
Title
}
