% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-pattern-recognition.R
\name{plotPrAssessment}
\alias{plotPrAssessment}
\title{Plot pattern related}
\usage{
plotPrAssessment(object, of_sample = NA)
}
\description{
Blueprint - currently not reliably working.
}
