% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-pattern-recognition.R
\name{plotPrSummary}
\alias{plotPrSummary}
\title{Title}
\usage{
plotPrSummary(
  object,
  of_sample = "",
  plot_type = "density_2d",
  display_points = FALSE,
  pt_size = 1,
  clrp = "milo"
)
}
\arguments{
\item{clrp}{}
}
\value{

}
\description{
Title
}
