% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_miscellaneous.R
\name{plotPseudotime}
\alias{plotPseudotime}
\title{Monocle3 Pseudotime}
\usage{
plotPseudotime(
  object,
  use_cds_file = FALSE,
  save_cds_file = FALSE,
  preprocess_method = "PCA",
  cluster_method = c("leiden", "louvain"),
  color_to = "pseudotime",
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{use_cds_file}{A directory leading to a .rds file containing a valid
cell_data_set-object previously calculated for the specified object. Specified
as a character value. If set to FALSE the cell_data_set object will be created
from scratch.}

\item{save_cds_file}{A filename/directory (that does not already exists) under which the used or created cell_data_set-object
is going to be stored specified as a character value. Should end with .rds.}

\item{preprocess_method}{Given to \code{monocle3::preprocess_cds()} if \code{use_cds_file} isn't a character string.}

\item{cluster_method}{Given to \code{monocle3::cluster_cells()} if \code{use_cds_file} isn't a character string.}

\item{verbose}{Logical value. If set to TRUE informative messages with respect
to the computational progress made will be printed.

(Warning messages will always be printed.)}

\item{feature_name}{The name under which the created pseudotime-variable is stored in the provided object. Will overwrite
already existing features of the same name!}
}
\value{
Returns a list of two ggplot-objects that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Monocle3 Pseudotime
}
