% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotScatter}
\alias{plotScatter}
\title{Title}
\usage{
plotScatter(
  object,
  of_sample = "",
  variables,
  pt_size = 1.5,
  pt_alpha = 0.8,
  pt_clr = "black",
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
