% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plot segmentation}
\usage{
plotSegmentation(
  object,
  encircle = TRUE,
  pt_size = NULL,
  pt_clrp = NULL,
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{encircle}{Logical. If set to TRUE the segments are enclosed in a polygon.}

\item{pt_size}{Numeric value. Specifies the size of points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\description{
Displays the segmentation of a specified sample that was drawn with
\code{SPATA::createSegmentation()}.
}
