% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotStatistiscInteractive}
\alias{plotStatistiscInteractive}
\title{Title}
\usage{
plotStatistiscInteractive(spata_df)
}
\arguments{
\item{spata_df}{}
}
\value{

}
\description{
Title
}
\examples{

}
