% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surface_family.R
\name{plotSurface}
\alias{plotSurface}
\alias{plotSurface2}
\alias{plotSurfaceInteractive}
\title{Plot the sample}
\usage{
plotSurface(
  data,
  color_to,
  pt_size = 2,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  pt_clrsp_dir = 1,
  image = NULL
)

plotSurface2(
  object,
  of_sample,
  color_to,
  method_gs = "mean",
  smooth = FALSE,
  smooth_span = 0.02,
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  display_image = F,
  display_title = F,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)

plotSurfaceInteractive(object)
}
\arguments{
\item{data}{A data.frame containing at least the variables \emph{'x'} and \emph{'y'}.}

\item{color_to}{The information to be displayed by color specified as a
character vector. If you specify a feature or a gene set this vector needs
to be of length one. If you specify more than one gene the average
expression of these genes will be calculated.}

\item{pt_size}{The size of the points specified as a numeric value.}

\item{pt_alpha}{The transparency of the points specified as a numeric value.}

\item{pt_clrsp}{The colour spectrum used to display \code{color_to} if the
specified variable is continuous. Needs to be one of \emph{'inferno', 'magma',
'plasma', 'cividis' or 'viridis'}.}

\item{pt_clrsp_dir}{The direction of the color spectrum specified as either \emph{1}
or \emph{-1}.}

\item{image}{An image of class \emph{Image} to be displayed in the background.
Easily accessible via \code{SPATA::image()}.}

\item{object}{A valid object of class \emph{spata}.}

\item{of_sample}{The samples name specified as a character of length one.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}. Ignored if \code{color_to} isn't a gene set.}

\item{smooth}{Logical value. If set to TRUE \code{plotSurface()} will smooth
the values displayed by color deploying \code{stats::loess()}.}

\item{smooth_span}{Numeric value, given to \code{stats::loess()} if
\code{smooth} is set to TRUE.}

\item{display_image}{Logical value to specify whether the histology image
of the sample is supposed to be displayed underneath the plot.

(Can be made visible with low \code{pt_alpha} or \code{color_to} set to NULL.)}

\item{display_title}{Logical value to specify whether a title is supposed to
be displayed.}

\item{assign}{Logical value. If set to TRUE a list of the data needed to plot the
returned plot will be assigned to the global environment.}

\item{assign_name}{The name of the list that is assigned to the global environment
specified as a character value that does not already exist.}

\item{verbose}{Logical value to specify whether informative messages are
supposed to be printed or not.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the sample and colors the surface according to the
expression of genes and gene sets or other featured characteristics.

\itemize{
 \item{ \code{plotSurface()} Takes a data.frame as the starting point.}
 \item{ \code{plotSurface2()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional paramters.}
 \item{ \code{plotSurfaceInteractive()} Takes only the spata-object and opens a mini-shiny
 application which allows for interactive plotting.}

}
}
