% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{plotSurfaceQuantile}
\alias{plotSurfaceQuantile}
\title{Title}
\usage{
plotSurfaceQuantile(
  object,
  color_to,
  pt_size = 2,
  smooth = TRUE,
  smooth_span = 0.02,
  n_quantiles = 5,
  mark_below = 5,
  ...
)
}
\arguments{
\item{pt_size}{}
}
\value{

}
\description{
Title
}
\examples{

}
