% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurfaceQuantiles}
\alias{plotSurfaceQuantiles}
\title{Title}
\usage{
plotSurfaceQuantiles(
  object,
  of_sample = "",
  color_to,
  n_qntls = 5,
  keep_qntls = 1:n_qntls,
  pt_alpha = 0.9,
  pt_clrp = "milo",
  pt_size = 2,
  smooth = TRUE,
  smooth_span = 0.02
)
}
\arguments{
\item{smooth_span}{}
}
\value{

}
\description{
Title
}
