% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectory}
\alias{plotTrajectory}
\title{Plot trajectory}
\usage{
plotTrajectory(
  object,
  trajectory_name,
  of_sample = "",
  color_to = NULL,
  method_gs = "mean",
  smooth = FALSE,
  smooth_span = 0.02,
  pt_size = 2.5,
  pt_alpha = 0.4,
  pt_clr = "red",
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  sgmt_size = 1,
  display_image = TRUE,
  display_title = FALSE,
  uniform_genes = "discard",
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{color_to}{The variable to be displayed by color:

 \itemize{
  \item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} as a character vector. If more than one gene is specified the average
  expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
  \item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
  }}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{pt_size}{The size of the points specified as a single numeric value.}

\item{pt_alpha}{The transparency of the points specified as single numeric value.}

\item{pt_clr}{The base color of every point displayed in the plot.}

\item{pt_clrp}{The color panel to be used if the specified variable displayed by
color is categorical/discrete. Run \code{all_colorpanels()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{all_colorspectra()} to see valid input..}

\item{sgmt_size}{The size of the segment arrrow specified as a numeric value.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}

\item{display_title}{Logical. If set to TRUE an informative title is displayed.}

\item{uniform_genes}{Character value. If set to \emph{'discard'} genes that are
uniformly expressed across all barcode-spots of the specified coordinates
data.frame are discarded. If set to \emph{'keep'} they are kept.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays a trajectory of a specified sample that was
drawn with \code{SPATA::createTrajectories()}.
}
