% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryAssessment}
\alias{plotTrajectoryAssessment}
\title{Trajectory trends distribution}
\usage{
plotTrajectoryAssessment(
  atdf,
  limits = c(NA, NA),
  plot_type = "density",
  binwidth = 1
)
}
\arguments{
\item{atdf}{An assessed trajectory data.frame returned by
\code{assessTrajectoryTrends()}.}

\item{limits}{The minimum and maximum auc-values to include. Given to
\code{ggplot2::scale_x_continuous()}.}

\item{plot_type}{One of \emph{'histogram', 'density', and 'ridgeplot'}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Visualizes the distribution of the assessment-scores
 (residuals-area-under-the-curve) of a trajectory via histograms.
}
