% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryFeatures}
\alias{plotTrajectoryFeatures}
\alias{plotTrajectoryGenes}
\alias{plotTrajectoryGeneSets}
\title{Trajectory line plots}
\usage{
plotTrajectoryFeatures(
  object,
  trajectory_name,
  of_sample,
  features = "percent.mt",
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = TRUE,
  verbose = TRUE
)

plotTrajectoryGenes(
  object,
  trajectory_name,
  of_sample,
  genes,
  average_genes = FALSE,
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = T,
  verbose = T
)

plotTrajectoryGeneSets(
  object,
  trajectory_name,
  of_sample,
  gene_sets = "Neftel_NPC_Comb",
  method_gs = "mean",
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}

\item{features}{The features of interest specified as a character vector.}

\item{smooth_method}{The smoothing method that will be used specified as a
single character value (e.g. \emph{"lm", "glm", "gam", "loess"}).}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{smooth_se}{Logical. If set to TRUE the confidence interval will be
displayed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{genes}{The genes of interest specified as a character vector.}

\item{average_genes}{Logical value. If set to TRUE the mean expression
of all specified genes will be calculated.}

\item{gene_sets}{The gene sets of interest specified as a character vector.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays values along a trajectory direction with
a smoothed lineplot.
}
