% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryHeatmap}
\alias{plotTrajectoryHeatmap}
\title{Expression dynamic in heatmap}
\usage{
plotTrajectoryHeatmap(
  object,
  of_sample,
  trajectory_name,
  variables = NULL,
  method_gs = "mean",
  show_row_names = TRUE,
  arrange_rows = "none",
  verbose = TRUE,
  split_columns = TRUE,
  borders = TRUE,
  smooth_span = 0.5,
  clrsp = "inferno"
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{variables}{The variables of interest specified as a character vector:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 }

All elements of the specified character vector must either belong to
gene sets or to genes.}

\item{show_row_names}{Logical. If set to TRUE the variable elements
will be displayed at the rownames of the heatmap.}

\item{arrange_rows}{Alter the way the rows of the heatmap
are displayed in order to highlight patterns. Currently either \emph{'maxima'}
or \emph{'minima'}.

Given to \code{accross} of \code{confuns::arrange_rows()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{split_columns}{Logial. If set to TRUE the heatmap is vertically
splitted according to the trajectory parts.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{clrsp}{The color spectrum to be used. Needs to be one of \emph{'inferno', 'magma', 'plasma', 'cividis' or 'viridis'}.}
}
\value{
A drawn heatmap.
}
\description{
Displays gene- or gene-set-expression values along a trajectory
direction with a smoothed heatmap.
}
