% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotUMAP}
\alias{plotUMAP}
\alias{plotTSNE}
\title{Plot dimensional reduction}
\usage{
plotUMAP(
  object,
  of_sample,
  color_to = NULL,
  method_gs = "mean",
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  verbose = TRUE
)

plotTSNE(
  object,
  of_sample,
  color_to = NULL,
  method_gs = "mean",
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) name specified as a character.}

\item{color_to}{The information you want to display by color specified as a
character vector. If you specify a feature or a gene set this vector needs
to be of length one. If you specify more than one gene the average
expression of these genes will be calculated and displayed by color.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{pt_size}{The size of the points specified as a numeric value.}

\item{pt_alpha}{The transparency of the points specified as a numeric value.}

\item{pt_clrsp}{The colour spectrum used to display \code{color_to} if the
specified variable is continuous. Needs to be one of \emph{inferno, magma,
plasma, cividis or viridis}.}

\item{verbose}{Logical value. If set to TRUE informative messages with respect
to the computational progress made will be printed.

(Warning messages will always be printed.)}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Plot dimensional reduction
}
