% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-family.R
\name{printDefaultInstructions}
\alias{printDefaultInstructions}
\title{Print current default settings}
\usage{
printDefaultInstructions(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Object of class glue.
}
\description{
Print current default settings
}
