% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-analysis.R
\name{rankTrajectoryTrends}
\alias{rankTrajectoryTrends}
\title{Trajectory expression analysis}
\usage{
rankTrajectoryTrends(stdf)
}
\arguments{
\item{stdf}{A summarized trajectory data.frame. (e.g. obtained by
\code{getSummarizedTrajectoryDf()}).}
}
\value{
A nested data.frame with information about the dynamics of each gene
or gene set.
}
\description{
Analyze the expression dynamics along
a specified trajectory by fitting a variety of models to the genes or
gene sets expression trends.
}
