% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-denoising.R
\name{runAutoencoderDenoising}
\alias{runAutoencoderDenoising}
\title{Denoise expression matrix}
\usage{
runAutoencoderDenoising(
  object,
  activation,
  bottleneck,
  layers = c(128, 64, 32),
  dropout = 0.1,
  epochs = 20,
  plot = FALSE,
  genes,
  discard_scaled_mtr = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{activation}{Character value. Denotes the activation function. (default = \emph{'relu'})}

\item{bottleneck}{Numeric value. Denotes the number of bottleneck neurons. (default = 32)}

\item{layers}{Numeric vector of length 3. Denotes the number of neurons in the three hidden layers.
(default = c(128, 64, 32))}

\item{dropout}{Numeric value. Denotes the dropout. (default = 0.1)}

\item{epochs}{Numeric value. Denotes the epochs of the neural network. (default = 20)}

\item{plot}{Logical. If set to TRUE a scatter plot of the result is displayed in the viewer pane.
See documentation for \code{plotAutoencoderResults()} for more information.}

\item{genes}{Character vector of length two. Denotes the genes to be used for the validation plot.}

\item{discard_scaled_mtr}{Logical. If set to TRUE the matrix in slot @data$scaled is discarded.}
}
\value{
A spata-object containing the denoised expression matrix in slot @data$denoised. This matrix
is then denoted as the active matrix.
}
\description{
This function constructs and uses a neural network to denoise
expression levels spatially.
}
