% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-main.R
\name{runPatternRecognition}
\alias{runPatternRecognition}
\title{Title}
\usage{
runPatternRecognition(
  object,
  of_sample = "",
  genes = 1500,
  genes_additional = NULL,
  with_ties = TRUE,
  threshold_qntl = 0.75,
  threshold_stw = 0.5,
  threshold_stpv = 0.1,
  smooth_span = 0.02,
  max_patterns = 9,
  n_start = 15,
  verbose = TRUE
)
}
\arguments{
\item{n_start}{}
}
\value{

}
\description{
Title
}
