% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{runPca}
\alias{runPca}
\alias{runPca2}
\title{Run Principal Component Analysis}
\usage{
runPca(object, of_sample = "", pca_comp = 30, mtr_name = NULL)

runPca2(object, of_sample = "", pca_comp = 30, mtr_name = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{pca_comp}{Numeric value. Denotes the number of principal components to be computed.}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}
}
\description{
Takes the expression matrix of choice and passes it to
\code{irlba::prcomp_irlba()}.
}
