% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{runTsne}
\alias{runTsne}
\alias{runTsne2}
\title{Run t-Stochastic Neighbour Embedding}
\usage{
runTsne(object, of_sample = "", tsne_perplexity = 30)

runTsne2(object, of_sample = "", tsne_perplexity = 30)
}
\arguments{
\item{tsne_perplexity}{}
}
\value{

}
\description{
Run t-Stochastic Neighbour Embedding
}
