% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{saveSpataObject}
\alias{saveSpataObject}
\alias{saveCorrespondingCDS}
\alias{saveCorrespondingSeuratObject}
\title{Save objects}
\usage{
saveSpataObject(object, new_directory = NULL, combine_with_wd = "/")

saveCorrespondingCDS(cds, object, new_directory = NULL, combine_with_wd = "/")

saveCorrespondingSeuratObject(
  seurat_object,
  object,
  new_directory = NULL,
  combine_with_wd = "/"
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{new_directory}{Character value. The new directory under which
to store the object of interest. Overwrites the stored default directory.
Use \code{getDefaultDirectory()} to obtain the current set up.}

\item{combine_with_wd}{Character value or FALSE. If specified with a
character value (default: \emph{'/'}) the input of \code{new_directory}
is considered to be a relative directory and is combined with the
current working directory (\code{base::getwd()}) with the character string
specified as the separator. If set to FALSE the input of \code{new_directory}
is taken as is.}

\item{cds}{A valid cell-data-set. (from the monocle3 platform)}

\item{seurat_object}{A valid seurat-object. (from the Seurat platform)}
}
\value{
Apart from their side effect (saving the object of interest) the functions return:
 \itemize{
 \item{ \code{saveSpataObject()}: An invisible TRUE. }
 \item{ \code{saveCorrespondingCDS(), saveCorrespondingSeuratObject()}: An updated spata-object that now contains the
 newest directory under which the corresponding cell-data-set or seurat-object has
 been saved. }
 }
}
\description{
Family of functions to save corresponding objects of different analysis
platforms. See details and value for more information.
}
\details{
If \code{new_directory} is set to NULL (the default) all functions first check if the spata-object contains any
deposited default directories. If so the specified object to be saved is saved under
that direction. If argument \code{directory} is specified it's input is taken as the
directory under which to store the object and the deposited directory is overwritten
such that the next time you load the spata-object it contains the updated directory.
In order for that to workthe \code{saveCorresponding*()}-functions return the
updated spata-object apart from saving the object of interest while \code{saveSpataObject()} simply returns an invisible TRUE
as the directory is stored inside the object before it is saved.
}
