% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_documentation.R
\docType{class}
\name{scvelo-class}
\alias{scvelo-class}
\alias{scvelo}
\title{scvelo object}
\value{
S4 object
}
\description{
scvelo object
}
\section{Slots}{

\describe{
\item{\code{scvelo}}{data.frame.}

\item{\code{scvelo1}}{data.frame.}
}}

