% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setActiveExpressionMatrix}
\alias{setActiveExpressionMatrix}
\title{}
\usage{
setActiveExpressionMatrix(object, of_sample = "", mtr_name)
}
\arguments{
\item{object}{A valid spata-object.}

\item{name}{Character value. The name of the matrix that is to be set as
the active expression matrix.}
}
\value{

}
\description{

}
