% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setActiveMatrix}
\alias{setActiveMatrix}
\title{}
\usage{
setActiveMatrix(object, name)
}
\arguments{
\item{name}{Character value. The name of the matrix that is to be set as
the active expression matrix.}
}
\value{

}
\description{

}
