% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setAutoencoderAssessment}
\alias{setAutoencoderAssessment}
\title{Title}
\usage{
setAutoencoderAssessment(object, assessment_list, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{assessment_list}{Named list with slots \code{$df} and \code{$set_up}.}
}
\value{
A spata-object.
}
\description{
Title
}
