% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setAutoencoderInfo}
\alias{setAutoencoderInfo}
\title{Title}
\usage{
setAutoencoderInfo(object, info_list)
}
\arguments{
\item{object}{A valid spata-object.}

\item{info_list}{A named list with slots \code{$activation, $assesment, $bottleneck, $dropout, $epochs, $layers}.}
}
\value{
A spata-object.
}
\description{
Title
}
