% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setDefaultInstructions}
\alias{setDefaultInstructions}
\alias{setDirectoryInstructions}
\title{Title}
\usage{
setDefaultInstructions(object)

setDirectoryInstructions(object)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
