% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setSpCorResults}
\alias{setSpCorResults}
\title{Title}
\usage{
setSpCorResults(object, sp_cor_list, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{assessment_list}{}
}
\value{

}
\description{
Title
}
