% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_server.R
\name{shiny_server_spatial_segmentation}
\alias{shiny_server_spatial_segmentation}
\alias{shiny_server_spatial_trajectories}
\title{Shiny server functions}
\usage{
shiny_server_spatial_segmentation(input, output, session)

shiny_server_spatial_trajectories(input, output, session)
}
\arguments{
\item{input}{shiny-input}

\item{output}{shiny-output}

\item{session}{shiny-session}
}
\description{
Shiny server functions
}
