% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_uis.R
\name{shiny_ui_spatial_segmentation}
\alias{shiny_ui_spatial_segmentation}
\alias{shiny_ui_spatial_trajectories}
\title{Shiny user interfaces}
\usage{
shiny_ui_spatial_segmentation()

shiny_ui_spatial_trajectories()
}
\value{
A shiny user interface
}
\description{
Shiny user interfaces
}
