% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{spata-class}
\alias{spata-class}
\alias{spata}
\title{spata object}
\value{
S4 object
}
\description{
spata object
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{data.frame.}

\item{\code{data}}{data_counts.}

\item{\code{dim_red}}{dim_red.}

\item{\code{fdata}}{data.frame.}

\item{\code{image}}{list.}

\item{\code{samples}}{character.}

\item{\code{scvelo}}{scvelo.}

\item{\code{trajectories}}{list.}

\item{\code{used_genesets}}{data.frame.}

\item{\code{version}}{list.}
}}

