% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\docType{data}
\name{spata_version}
\alias{spata_version}
\title{The current spata-version}
\format{An object of class \code{list} of length 4.}
\usage{
spata_version
}
\description{
The current spata-version
}
\keyword{datasets}
