% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{spatial_trajectory-class}
\alias{spatial_trajectory-class}
\alias{spatial_trajectory}
\title{spatial_trajectory object}
\value{
S4 object
}
\description{
spatial_trajectory object
}
\section{Slots}{

\describe{
\item{\code{compiled_trajectory_df}}{data.frame.}

\item{\code{segment_trajectory_df}}{data.frame.}

\item{\code{ranked_genes_df}}{data.frame.}

\item{\code{comment}}{character.}

\item{\code{name}}{character.}

\item{\code{sample}}{character.}
}}

