% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{spatialTrajectory-class}
\alias{spatialTrajectory-class}
\alias{spatialTrajectory}
\title{spatialTrajectory object}
\value{
S4 object
}
\description{
spatialTrajectory object
}
\section{Slots}{

\describe{
\item{\code{compiled_trajectory_df}}{data.frame.}

\item{\code{segment_trajectory_df}}{data.frame.}

\item{\code{ranked_genes_df}}{data.frame.}

\item{\code{comment}}{character.}

\item{\code{name}}{character.}

\item{\code{sample}}{character.}
}}

