% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spata-object-manipulation.R
\name{subsetBySample}
\alias{subsetBySample}
\title{Subset a spata-object}
\usage{
subsetBySample(object, sample)
}
\arguments{
\item{object}{A valid spata-object.}

\item{sample}{Character value. The sample of interest.}
}
\value{
An updated spata-object.
}
\description{
Reduces the number of samples in the
specified spata-object to one.
}
