% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spata-object-manipulation.R
\name{subsetSpataObject}
\alias{subsetSpataObject}
\title{Reduce the number of samples}
\usage{
subsetSpataObject(object, sample)
}
\arguments{
\item{object}{A valid spata-object.}

\item{sample}{Character value. The sample of interest.}
}
\value{
An updated spata-object.
}
\description{
Keeps only information of the specified sample
}
