% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-analysis.R
\name{trajectoryTrend}
\alias{trajectoryTrend}
\title{Trajectory trends}
\usage{
trajectoryTrend(atdf, limit = 2, pattern)
}
\arguments{
\item{atdf}{An assessed trajectory data.frame.}

\item{limit}{Numeric value. The maximum area-under-the-curve value the
trajectory-trend-assessment might have.}

\item{pattern}{Character vector. The patterns of interest.}
}
\value{
A character vector of gene or gene-set names that follow the specified
patterns to the specified degree.
}
\description{
Extracts the trajectories of a desired trend.
}
