% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-family.R
\name{transformSeuratToSpata}
\alias{transformSeuratToSpata}
\title{Transform seurat-object to spata-object}
\usage{
transformSeuratToSpata(
  seurat_object,
  sample_name,
  assay = "Spatial",
  method = "spatial",
  coords_from = "umap",
  gene_set_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{A valid seurat object.}

\item{sample_name}{Character value. Future input for SPATA's \code{of_sample}-argument.}

\item{method}{Character value. Determines the data slots from which to compile the spata-object.

 \describe{
  \item{\emph{'spatial'}}{Denotes that the data to be used derived from spatial experiments. }
  \item{\emph{'single_cell'}}{Denotes that the data to be used derived from single cell experiments. }
 }}

\item{coords_from}{Character value. Either \emph{'umap'} or \emph{'tsne'}.

 Only relevant if \code{method} was set to \emph{'single_cell'}. Denotes the slot from which to
 take the surrogate coordinates.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A spata object.
}
\description{
Transform seurat-object to spata-object
}
