% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{verbose}
\alias{verbose}
\title{verbose}
\usage{
verbose(verbose)
}
\arguments{
\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
verbose
}
