% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_family.R
\name{addGeneSet}
\alias{addGeneSet}
\title{Add a new gene set}
\usage{
addGeneSet(object, gs_class, gs_name, gs_genes, overwrite = FALSE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{gs_class}{The class the gene set belongs to specified as a character.}

\item{gs_name}{The name of the new gene set specified as a character.}

\item{gs_genes}{The genes specified as a character vector.}

\item{overwrite}{Logical. Overwrites existing gene sets with the same \code{gs_class} -
\code{gs_name} combination.}
}
\value{
An updated spata-object.
}
\description{
Stores a new gene set in the spata-object.
}
\details{
Coerces \code{gs_class} and \code{gs_name} to the final gene set name.
}
