% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-denoising.R
\name{autoEncoderDenoising}
\alias{autoEncoderDenoising}
\title{Deep Autoencoder for Denoising}
\usage{
autoEncoderDenoising(
  object,
  layers = c(128, 64, 32),
  bottleneck = 16,
  dropout = 0.1,
  activation = "relu",
  epochs = 20,
  val_plot = FALSE,
  val_genes = c("CD163", "CD68"),
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{layers}{Numeric vector of length 3. Denotes the number of neurons in the three hidden layers.
(default = c(128, 64, 32))}

\item{bottleneck}{Numeric value. Denotes the number of bottleneck neurons. (default = 32)}

\item{dropout}{Numeric value. Denotes the dropout. (default = 0.1)}

\item{activation}{Character value. Denotes the activation function. (default = \emph{'relu'})}

\item{epochs}{Numeric value. Denotes the epochs of the neural network. (default = 20)}

\item{val_plot}{Logical. If set to TRUE a scatter plot of the result is displayed.}

\item{val_genes}{Character vector of length two. Denoting the genes to be used for the validation plot.}
}
\value{
A spata-object containing the denoised expression matrix.
}
\description{
This function constructs and uses a neural network to denoies
expression levels spatially.
}
