% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_family.R
\name{check_color_to}
\alias{check_color_to}
\alias{check_variables}
\title{Check input}
\usage{
check_color_to(
  color_to,
  all_features = character(),
  all_gene_sets = character(),
  all_genes = character(),
  max_length = 25
)

check_variables(
  variables,
  all_features = character(),
  all_gene_sets = character(),
  all_genes = character(),
  max_length = 25,
  simplify = FALSE
)
}
\arguments{
\item{color_to}{character or list}

\item{all_features}{character. all features}

\item{all_gene_sets}{character. all gene sets}

\item{all_genes}{character. all genes}
}
\value{
a named list (or a character vector if max_length is set to 1).
}
\description{
Checks what of the color_to input is actually given in the object.
The all_* parameters allow for manipulation of what color_to input is an
option. (e.g. if color_to of the specific function must not be of type 'features'
specify paramter 'all_features' as an empty character vector and no color_to element
can be recognized as such.)
}
