% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_family.R
\name{check_gene_sets}
\alias{check_gene_sets}
\title{Check input}
\usage{
check_gene_sets(object, gene_sets, max_length = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{gene_sets}{The gene_sets-input.}

\item{max_length}{The maximum length the input can have.}
}
\value{
- If the input is just fine: the input.
- If the input suffices for the function to work but needs slight moderation:
a modified version of the input along with a warning-message.
- If the input does not suffice for the function to work: an error
}
\description{
The \code{check_()}-function family checks a functions input to
make sure that the function can run bug-free. If the input does not suffice a
helping message is printed for the user to know what needs to be adjusted in
order to make the function work.
}
