% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_pattern}
\alias{check_pattern}
\title{Title}
\usage{
check_pattern(object, of_sample = "", patterns = "", method_pr = "hotspot")
}
\arguments{
\item{of_pattern}{}
}
\value{

}
\description{
Title
}
