% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords-family.R
\name{coordsSpatial}
\alias{coordsSpatial}
\alias{coordsSegment}
\alias{coordsTrajectory}
\alias{coordsDimRed}
\alias{coordsUMAP}
\alias{coordsTSNE}
\title{Extract coordinates}
\usage{
coordsSpatial(object, of_sample)

coordsSegment(object, of_sample, segment_name)

coordsTrajectory(object, of_sample, trajectory_name)

coordsDimRed(object, of_sample, method_dr = c("UMAP", "TSNE"))

coordsUMAP(object, of_sample)

coordsTSNE(object, of_sample)
}
\arguments{
\item{object}{A valid object of class \emph{spata}.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}

\item{segment_name}{The segment of interest specified as a single character
value.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{method_dr}{The dimensional reduction method of
interest specified as a single character value. (Currently
either \emph{'UMAP'} or \emph{'TSNE'}).}
}
\value{
A data.frame that contains the unique identifiers
(keys): \emph{barcodes, sample} and the coordinates.

 \itemize{
  \item{ \code{coordsSpatial()}:  \emph{x, y}}
  \item{ \code{coordsSegment()}: \emph{x, y}}
  \item{ \code{coordsTrajectory()}: \emph{x, y}}
  \item{ \code{coordsTSNE()}: \emph{tsne1, tsne2}}
  \item{ \code{coordsUMAP()}: \emph{umap1, umap2}}
  }
}
\description{
Extracts spatial information of whole samples, trajectories,
segments etc. in form of coordinates. (See 'Value').
}
