% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{evaluate_gene_cluster_tendency}
\alias{evaluate_gene_cluster_tendency}
\alias{evaluate_gene_cluster_tendency_dbscan}
\alias{evaluate_gene_cluster_tendency_dbscan2}
\alias{evaluate_gene_cluster_tendency_hdbscan}
\title{Iterate over marked data.frame and evaluate cluster/pattern tendency}
\usage{
evaluate_gene_cluster_tendency(
  marked_df,
  verbose = TRUE,
  pb = NULL,
  ref_totwss,
  ...
)

evaluate_gene_cluster_tendency_dbscan(marked_df, verbose = TRUE, pb = NULL)

evaluate_gene_cluster_tendency_dbscan2(marked_df, verbose = TRUE, pb = NULL)

evaluate_gene_cluster_tendency_hdbscan(marked_df, verbose = TRUE, pb = NULL)
}
\arguments{
\item{marked_df}{A data.frame with x- and y-coordinates corresponding
to the barcode-spots that represent the underlying expression pattern.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{pb}{A progress_bar-object.}
}
\description{
Use within \code{purrr::map()} to iterate over
nested data.frame containing marked/dropped data.frames in
variable.
}
