% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-helper-functions.R
\name{examineDeResults}
\alias{examineDeResults}
\title{Examine differentially expression results}
\usage{
examineDeResults(de_df, clrp = "milo", return = "both", binwidth = 10, ...)
}
\arguments{
\item{return}{Character value. Denotes what to plot. One of \emph{'both', 'n_genes'}
and \emph{'d_logFC'}}

\item{...}{Additional parameters given to \code{ggplot2::facet_wrap()}}
}
\description{
Visualizes the number of differentially expressed genes
and/or the distribution of logFC-values per cluster.
}
