% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-expression-analysis.R
\name{findDE}
\alias{findDE}
\title{Find marker genes}
\usage{
findDE(object, of_sample, feature, method_de = "wilcox", p_val_adj = 0.05)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}

\item{feature}{Character value. A categorical feature of the feature data that
separates all barcodes into subgroups for which eventually the marker genes will be determined.

(Number of subgruops must be lower than 20.)}

\item{method_de}{Character value. Given to argument \code{test.use} of \code{Seurat::FindAllMarkers()}.}

\item{p_val_adj}{The maximum adjusted p-value allowed in the output.}
}
\value{
A DE-object.
}
\description{
Finds the differentially expressed genes of a set of subgroups.
}
