% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-additional.R
\name{findNPatterns}
\alias{findNPatterns}
\title{Title}
\usage{
findNPatterns(object, of_sample = "", n_patterns, plotPrSummary = TRUE)
}
\arguments{
\item{plotPrSummary}{}
}
\value{

}
\description{
Title
}
