% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-sub.R
\name{generate_reference_df}
\alias{generate_reference_df}
\alias{generate_sample}
\title{Generates a reference data.frame}
\usage{
generate_reference_df(n, x_coords, y_coords)

generate_sample(x, l = 250)
}
\description{
Generates a reference data.frame
}
