% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getActiveMatrixName}
\alias{getActiveMatrixName}
\title{Obtain name of currently active matrix}
\usage{
getActiveMatrixName(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Character value.
}
\description{
Obtain name of currently active matrix
}
