% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getBarcodes}
\alias{getBarcodes}
\title{Obtain barcodes of a sample}
\usage{
getBarcodes(object, of_sample = "all")
}
\arguments{
\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}
}
\value{
All barcodes of the specified sample(s).
}
\description{
Obtain barcodes of a sample
}
