% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getBarcodes}
\alias{getBarcodes}
\title{Obtain barcodes of a sample}
\usage{
getBarcodes(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
All barcodes of the specified sample(s) as a character vector.
}
\description{
Obtain barcodes of a sample
}
