% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getDefaultInstructions}
\alias{getDefaultInstructions}
\alias{getDefaultDirectory}
\title{Title}
\usage{
getDefaultInstructions(object)

getDefaultDirectory(object, to)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Title
}
