% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\title{Obtain expression matrix}
\usage{
getExpressionMatrix(object, of_sample = "all")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}
}
\value{
The expression matrix of the specified object and sample.
}
\description{
Obtain expression matrix
}
