% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\alias{getCountMatrix}
\title{Obtain count and expression matrix}
\usage{
getExpressionMatrix(object, of_sample = "", name = NULL)

getCountMatrix(object, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
The active expression or count matrix of the specified object and sample(s).
}
\description{
Obtain count and expression matrix
}
