% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGenes}
\alias{getGenes}
\alias{getGenesInteractive}
\title{Obtain gene names}
\usage{
getGenes(object, of_gene_sets = "all", in_sample = "",
  simplify = TRUE)

getGenesInteractive(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_gene_sets}{A character vector specifying the gene sets from which to
return the gene names.}

\item{in_sample}{The sample(s) in which the genes have to be expressed in order
to be included.}

\item{simplify}{Logical. If set to TRUE the list to be returned will be simplified
into a character vector.}
}
\value{
A list named according to the input of \code{of_gene_sets} in which each element is
a character vector containing the names of genes the specific gene set is
composed of. Is coalesced to a vector if \code{simplify} is set to TRUE.
}
\description{
Obtain gene names
}
