% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGenes}
\alias{getGenes}
\alias{getSpCorGenes}
\alias{getGenesInteractive}
\title{Obtain gene names}
\usage{
getGenes(
  object,
  of_gene_sets = NULL,
  of_hotspots = NULL,
  of_sample = "",
  in_sample = "",
  simplify = TRUE
)

getSpCorGenes(
  object,
  of_sample = "",
  similar_to = NULL,
  distinct_to = NULL,
  top_n = 25,
  method_hclust = NULL,
  cluster_names = NULL,
  simplify = TRUE
)

getGenesInteractive(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_gene_sets}{A character vector specifying the gene sets from which to
return the gene names.}

\item{in_sample}{The sample(s) in which the genes have to be expressed in order
to be included.}

\item{simplify}{Logical. If set to TRUE the list to be returned will be simplified
into a character vector.}

\item{of_pattern}{A character vector specifiying the patterns from which to return
the gene names. If denoted as \emph{""} the genes of all patterns are returned.
Set \code{simplify} to FALSE in order to return a named list.}
}
\value{
A list named according to the input of \code{of_gene_sets} in which each element is
a character vector containing the names of genes the specific gene set is
composed of. Is coalesced to a vector if \code{simplify} is set to TRUE.
}
\description{
Obtain gene names
}
