% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getHotspotInfoDf}
\alias{getHotspotInfoDf}
\alias{getHotspotGeneDf}
\title{Title}
\usage{
getHotspotInfoDf(object, of_sample = NA)

getHotspotGeneDf(object, of_sample = NA)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
