% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getPrResults}
\alias{getPrResults}
\alias{getPrSuggestion}
\alias{getPatternNames}
\title{Title}
\usage{
getPrResults(object, method_pr = "hotspot", of_sample = NA)

getPrSuggestion(object, method_pr = "hotspot", of_sample = NA)

getPatternNames(object, method_pr = "hotspot", of_sample = NA)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
