% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getPrResults}
\alias{getPrResults}
\alias{getPrSuggestion}
\alias{getPatternNames}
\title{Title}
\usage{
getPrResults(object, of_sample = "")

getPrSuggestion(object, of_sample = "")

getPatternNames(object, of_sample = "")
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
