% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{getSegmentNames}
\alias{getSegmentNames}
\title{Obtain segment names}
\usage{
getSegmentNames(object, of_sample)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The samples from which to obtain the segment names specified
as a character vector.}
}
\value{
A list named according to the \code{of_sample} in which each element is
a character vector containing the names of segments which were drawn for the
specific sample. Is simplified to a vector if the number of samples in the returned list
is one.
}
\description{
Obtain segment names
}
