% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{getTrajectoryNames}
\alias{getTrajectoryNames}
\title{Obtain trajectory names}
\usage{
getTrajectoryNames(object, of_sample = "all")
}
\arguments{
\item{object}{A valid spata object.}

\item{of_sample}{The samples from which to obtain the trajectory names specified
as a character vector.}
}
\value{
A list named according to the \code{of_sample} in which each element is
a character vector containing the names of trajectories which were drawn for the
specific sample. Is simplified to a vector if the number of samples in the returned list
is one.
}
\description{
Obtain trajectory names
}
