% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_assess_pattern_results}
\alias{hlpr_assess_pattern_results}
\alias{hlpr_assess_hotspot_results}
\title{Title}
\usage{
hlpr_assess_pattern_results(
  object,
  cluster_eval_df,
  max_patterns = NULL,
  n_start = NULL,
  of_sample = NULL,
  smooth_span = NULL,
  threshold_stpv = NULL,
  threshold_stw = NULL,
  n_patterns = NULL,
  verbose = TRUE,
  ...
)

hlpr_assess_hotspot_results(
  object,
  cluster_eval_df,
  ignored_genes = NULL,
  of_sample = NULL,
  smooth_span = NULL,
  threshold_certainty = 0.2,
  threshold_qntl = NULL,
  threshold_stpv = NULL,
  threshold_stw = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
