% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern-analysis.R
\name{igraph_cluster}
\alias{igraph_cluster}
\alias{plot_igraph_heatmap}
\title{Clustering with igraph}
\usage{
igraph_cluster(cor.mtr, num.nn = 20)

plot_igraph_heatmap(cor.mtr, num.nn, color.nn)
}
\arguments{
\item{cor.mtr}{A correlation matrix.}

\item{num.nn}{Numeric value. The maximum number of nearest neighbours to compute.
The default value is set to the smaller of the number of columnns in data. Given
to \code{RANN::nn2()} as input for argument \code{k}.}

\item{color.nn}{Numeric value.}
}
\value{
Cluster results.
}
\description{
Clustering with igraph
}
