% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinWith-family.R
\name{joinWithFeatures}
\alias{joinWithFeatures}
\alias{joinWithGenes}
\alias{joinWithGeneSets}
\alias{joinWithVariables}
\title{Join barcodes with additional variables}
\usage{
joinWithFeatures(
  object,
  coords_df,
  features,
  smooth = FALSE,
  smooth_span = 0.02,
  verbose = TRUE
)

joinWithGenes(
  object,
  coords_df,
  genes,
  average_genes = FALSE,
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE
)

joinWithGeneSets(
  object,
  coords_df,
  gene_sets,
  method_gs = "mean",
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE
)

joinWithVariables(
  object,
  coords_df,
  variables,
  method_gs = "mean",
  average_genes = FALSE,
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{coords_df}{A data.frame that contains at least the character variables
\emph{samples} and \emph{barcodes}.}

\item{features}{The features of interest specified as a character vector.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{genes}{The genes of interest specified as a character vector.}

\item{gene_sets}{The gene sets of interest specified as a character vector.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{variables}{A named list. Names must contain one of \emph{features, genes}
or \emph{gene_sets}. The input of \code{coords_df} will be joined with
the respective elements of each slot by calling the appropriate
\code{joinWith*()}-function.}
}
\value{
The input data.frame of \code{coords_df} joined with all the
specified variable-elements (by the key-variable \emph{barcodes}).
}
\description{
Each member of the \code{joinWith()}-family takes a data.frame as
input that contains at least the variables \emph{barcodes} and \emph{sample}.
(Easily obtained with the \code{spata::coords()}-family.) It joins them over
 the \emph{barcodes}-variable with a specified aspect.
}
