% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{loadSpataObject}
\alias{loadSpataObject}
\alias{saveSpataObject}
\title{Load and save a spata-object}
\usage{
loadSpataObject(input_path)

saveSpataObject(object, output_path, object_name, overwrite = FALSE)
}
\arguments{
\item{input_path}{Character value. The directory leading to the spata-object.}

\item{object}{A valid spata-object.}

\item{output_path}{Character value. Specifies the folder in which to store
the object.}

\item{object_name}{Character value. The name-suffix for the file name under which the
spata-object is stored. Is prefixed with \emph{'spata-obj-'}.}

\item{overwrite}{Logical. Needs to be set to TRUE if the resulting directory from
\code{output_path} and \code{object_name} already exists.}
}
\description{
Wrapper around \code{base::readRDS()} and \code{base::saveRDS()}.
}
