% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{loadSpataObject}
\alias{loadSpataObject}
\alias{saveSpataObject}
\title{Load and save a spata-object}
\usage{
loadSpataObject(input_path, update = FALSE, verbose = TRUE)

saveSpataObject(object, output_path, file_name, overwrite = FALSE)
}
\arguments{
\item{input_path}{Character value. The directory leading to the spata-object.}

\item{update}{Logical. If set to TRUE \code{updateSpataObject()} is used to check
whether the loaded spata-object is of the most recent version - if it's not it is
updated.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{object}{A valid spata-object.}

\item{output_path}{Character vector or NULL. Specifies the folder in which to store
the object if the directory is valid. If set to NULL saving is skipped.}

\item{file_name}{Character value. The name-suffix for the file name under which the
spata-object is stored if \code{output_path} is a valid directory. Is prefixed with
 \emph{'spata-obj-'} and suffixed with \emph{'.RDS'}.}

\item{overwrite}{Logical. Needs to be set to TRUE if the resulting directory from
\code{output_path} and \code{file_name} already exists.}
}
\description{
Wrapper around \code{base::readRDS()} and \code{base::saveRDS()}.
}
