% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotAutoencoderResults}
\alias{plotAutoencoderResults}
\title{Plot scaled vs. denoised expression}
\usage{
plotAutoencoderResults(
  object,
  genes,
  mtr_name = "denoised",
  scales = "free",
  pt_size = 1.5,
  pt_alpha = 0.75,
  pt_clrp = "milo",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{genes}{Character vector of length two. Denotes the genes to be used for the validation plot.}

\item{pt_size}{The size of the points specified as a single numeric value.}

\item{pt_alpha}{The transparency of the points specified as single numeric value.}

\item{pt_clrp}{The color panel to be used if the specified variable displayed by
color is categorical/discrete. Run \code{all_colorpanels()} to see valid input.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Compares the distribution of the expression levels of \code{genes}
between the scaled matrix and the denoised matrix.
}
\details{
This function requires a denoised matrix in slot @data generated
by \code{runAutoEncoderDenoising()} as well as a scaled matrix.
}
