% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDeHeatmap}
\alias{plotDeHeatmap}
\title{Plot differentially expressed genes}
\usage{
plotDeHeatmap(
  object,
  of_sample = "",
  across,
  across_subset = NULL,
  method_de = "wilcox",
  max_adj_pval = 0.5,
  n_highest_lfc = 50,
  n_lowest_pval = 50,
  breaks = NULL,
  genes = NULL,
  n_barcode_spots = NULL,
  hm_clrp = "milo",
  hm_colors = viridis::viridis(15),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{across}{Character value. The name of the discrete feature-variable of interest.

Hint: Character- or factor- variables are discrete variables. These functionally assign the barcode spots to distinct
groups or clusters (e.g. \emph{segment} or \emph{seurat_clusters}) whoose properties you might want
to compare against each other. Use \code{getFeatureNames()} to get an overview of the
features variables your spata-object contains.}

\item{across_subset}{Character vector or NULL. Specify the particular groups or clusters of interest the feature-variable
specified in argument \code{across} contains. If set to NULL all of them are chosen.

Hint: Use \code{getFeatureValues()} to obtain all available groups of a certain
feature-variable.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{max_adj_pval}{Numeric value. Sets the maximal threshold for adjusted p-values allowed. All genes
with adjusted p-values above that threshold are ignored.}

\item{n_highest_lfc}{Numeric value. Affects the total number of genes that are kept. See details.}

\item{n_lowest_pval}{Numeric value. Affects the total number of genes that are kept. See details.}

\item{breaks}{Denotes the colorspectrum breaks. If set to NULL the breaks are set automatically. If a
numeric vector is specified it is taken as input. If a function is specified the expression matrix is
passed to it as the first argument and the length of \code{hm_colors} as the second argument.}

\item{genes}{Character vector or NULL. If you want to display specific genes irrespective of de-anaylsis results you
can specifiy them in \code{genes}. If \code{genes} is specified that way arguments referring to de-anylsis results are
ignored and only the genes specified are taken and displayed.}

\item{n_barcode_spots}{The number of barcode-spots belonging to each cluster you want to
include in the matrix. Should be lower than the total number of barcode-spots of every cluster
and can be deployed in order to keep the heatmap clear and aesthetically pleasing.

If set to NULL (the default) it is automatically computed according to the number of genes that
are displayed in the heatmap.}

\item{hm_clrp}{The colorpanels used for the discrete heatmap annotation. Run \code{all_colorpanels()}
to obtain valid input options.}

\item{hm_colors}{A vector of colors to be used for the continuous heatmap annotation..}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional parameters given to \code{pheatmap::pheatmap()}.}
}
\value{
A heatmap of class 'pheatmap'.
}
\description{
Visualizes the expression of genes across subgroups in a heatmap. It either takes the results
from previously conducted de-analysis and uses the expression information of the spata-object to plot a heatmap displaying the
differently expressed genes of every group of the feature denoted in the argument \code{across}.
}
\details{
The de-data.frame is processed such that the following steps are performed for every experimental
group. (With "genes" we refer to the rows (observations) of \code{data}.)

\enumerate{
 \item{Discards genes with \emph{avg_logFC}-values that are either infinite or negative}
 \item{Discards genes with adjusted p-values above the threshold set with \code{max_adj_pval}}
 \item{Slices the data.frame in order that for every unique cluster of the \emph{cluster}-variable}:
 \enumerate{
  \item{the n genes with the highest \emph{avg_logFC}-values are kept where n = \code{n_highest_lfc}}
  \item{the n genes with the lowest \emph{p_val_adj}-values are kept where n = \code{n_lowest_pval}}
  }
 \item{Arranges the genes according to the highest \emph{avg_logFC}-values}
 }
}
