% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_miscellaneous.R
\name{plotDistribution}
\alias{plotDistribution}
\alias{plotDistribution2}
\alias{plotDistribution3}
\title{Visualize variable distribution}
\usage{
plotDistribution(
  data,
  variables = NULL,
  plot_type = "histogram",
  binwidth = 0.05,
  ...
)

plotDistribution2(
  object,
  of_sample,
  variables,
  method_gs = "mean",
  plot_type = "histogram",
  binwidth = 0.05,
  ...,
  normalize = TRUE,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)

plotDistribution3(
  object,
  of_sample,
  variables,
  across,
  method_gs = "mean",
  plot_type = "histogram",
  binwidth = 0.05,
  ...,
  normalize = TRUE,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)
}
\arguments{
\item{data}{The data.frame containing numeric variables.}

\item{variables}{The numeric variable whose distribution you want to display.}

\item{plot_type}{One of \emph{'histogram', 'density', 'violin', 'boxplot' and 'ridgeplot'}.}

\item{binwidth}{The binwidth to use if \code{plot_type} is specified as \emph{'histogram'}.}

\item{...}{additional arguments to \code{ggplot2::facet_wrap()}}

\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the data specified as a
character value.}

\item{verbose}{Logical value. If set to TRUE informative messages with respect
to the computational progress made will be printed.

(Warning messages will always be printed.)}

\item{across}{A categorical feature across which the value-distribution is displayed. Must be
a \emph{character or factor}-element of \code{SPATA::getFeatureNames()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Visualizes the distribution of values of a set of variables for the
whole sample accross specific subgroups.

\itemize{
 \item{ \code{plotDistribution()} Takes a data.frame as the starting point.}
 \item{ \code{plotDistribution2()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters.}
 \item{ \code{plotDistribution3()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters. It takes an additional
 argument which allows to display the variables value-distribution across subgroups.}
}
}
