% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotDistribution}
\alias{plotDistribution}
\alias{plotDistribution2}
\title{Visualize variable distribution}
\usage{
plotDistribution(
  data,
  variables = NULL,
  plot_type = "histogram",
  binwidth = 0.05,
  ...
)

plotDistribution2(
  object,
  of_sample,
  variables,
  method_gs = "mean",
  plot_type = "histogram",
  binwidth = 0.05,
  ...,
  normalize = TRUE,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)
}
\arguments{
\item{data}{The data.frame containing numeric variables.}

\item{variables}{The numeric variable whose distribution you want to display.}

\item{plot_type}{One of \emph{'histogram', 'density', 'violin', 'boxplot' and 'ridgeplot'}.}

\item{binwidth}{The binwidth to use if \code{plot_type} is specified as \emph{'histogram'}.}

\item{...}{additional arguments to \code{ggplot2::facet_wrap()}}

\item{object}{A valid object of class \emph{spata}.}

\item{of_sample}{The samples name specified as a character of length one.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}. Ignored if \code{color_to} isn't a gene set.}

\item{assign}{Logical value. If set to TRUE a list of the data needed to plot the
returned plot will be assigned to the global environment.}

\item{assign_name}{The name of the list that is assigned to the global environment
specified as a character value that does not already exist.}

\item{verbose}{Logical value to specify whether informative messages are
supposed to be printed or not.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Visualize variable distribution
}
