% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDistribution}
\alias{plotDistribution}
\alias{plotDistribution2}
\alias{plotDistributionAcross}
\alias{plotDistributionAcross2}
\title{Visualize variable distribution}
\usage{
plotDistribution(
  object,
  of_sample = "",
  variables,
  method_gs = "mean",
  plot_type = "histogram",
  binwidth = 0.05,
  ...,
  normalize = TRUE,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)

plotDistribution2(
  data,
  variables = "all",
  plot_type = "histogram",
  binwidth = 0.05,
  ...
)

plotDistributionAcross(
  object,
  of_sample = "",
  variables,
  across,
  method_gs = "mean",
  plot_type = "violin",
  binwidth = 0.05,
  ...,
  normalize = TRUE,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)

plotDistributionAcross2(
  data,
  variables = "all",
  across,
  plot_type = "violin",
  binwidth = 0.05,
  ...,
  normalize = TRUE,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{variables}{The numeric variables whose distribution you want to display. Specified as
a character vector.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{plot_type}{One of \emph{'histogram', 'density', 'violin', 'boxplot' and 'ridgeplot'}.}

\item{binwidth}{The binwidth to use if \code{plot_type} is specified as \emph{'histogram'}.}

\item{...}{additional arguments to \code{ggplot2::facet_wrap()}}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{assign}{Logical. If set to TRUE a named list will be assigned to the global
environment. This list contains data and information to rebuild or additionally
customize the output plot of this function.}

\item{assign_name}{The name the assigned list is supposed to have specified as
a single character value.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{data}{The data.frame containing numeric variables.}

\item{across}{A categorical feature across which the value-distribution is displayed. Must be
a \emph{character or factor}-element of \code{SPATA::getFeatureNames()}.}
}
\description{
Visualizes the distribution of values of a set of variables for the
whole sample across specific subgroups.

\itemize{
 \item{ \code{plotDistribution()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters.}
 \item{ \code{plotDistribution2()} Takes a data.frame as the starting point.}
 \item{ \code{plotDistributionAcross()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters. It takes an additional
 argument which allows to display the variables value-distribution across subgroups.}
}
}
