% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDistribution}
\alias{plotDistribution}
\alias{plotDistribution2}
\title{Distribution of continuous values}
\usage{
plotDistribution(
  object,
  variables,
  plot_type = "histogram",
  method_gs = NULL,
  clrp = NULL,
  binwidth = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotDistribution2(
  df,
  variables = "all",
  plot_type = "histogram",
  clrp = "milo",
  binwidth = 0.05,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{variables}{Character vector. The numeric variables of interest.}

\item{plot_type}{Character value. One of \emph{'histogram', 'density', 'violin', 'boxplot' and 'ridgeplot'}.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{binwidth}{The binwidth to use if \code{plot_type} is specified as \emph{'histogram'}.}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{additional arguments to \code{ggplot2::facet_wrap()}}

\item{df}{A data.frame that contains the numeric variables specified in \code{variables}.}
}
\description{
Visualizes the distribution of values of a set of variables.

\itemize{
 \item{ \code{plotDistribution()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters.}
 \item{ \code{plotDistribution2()} Takes a data.frame as the starting point.}
 }
}
