% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDistributionDiscrete}
\alias{plotDistributionDiscrete}
\title{Distribution of discrete features}
\usage{
plotDistributionDiscrete(
  object,
  of_sample = "",
  features,
  feature_compare = NULL,
  clrp = "milo",
  position = "fill",
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{features}{The features of interest specified as a character vector.}

\item{feature_compare}{Character vector or NULL. The discrete feature you want to compare the
features of \code{features} to.}

\item{clrp}{clrp params}

\item{position}{Character value. Given to \code{position} of \code{ggplot2::geom_bar()}. One of
\emph{'stack', 'dodge'} or \emph{'fill'}.}

\item{...}{Additional parameters given to \code{ggplot2::facet_wrap()}.}
}
\description{
Visualize the distribution of discrete features.
}
