% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-pattern-recognition.R
\name{plotGeneCenter}
\alias{plotGeneCenter}
\title{Title}
\usage{
plotGeneCenter(
  object,
  genes,
  plotSurfaceComparison = FALSE,
  pt_clr = NULL,
  pt_size = NULL,
  of_sample = NA
)
}
\arguments{
\item{plotSurface}{}
}
\value{

}
\description{
Title
}
