% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotGeneMetaData}
\alias{plotGeneMetaData}
\title{Title}
\usage{
plotGeneMetaData(
  object,
  variables,
  mtr_name = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_size = NULL,
  of_sample = NA
)
}
\arguments{
\item{mtr_name}{}
}
\value{

}
\description{
Title
}
