% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotHeatmapDE}
\alias{plotHeatmapDE}
\title{Plot differentially expressed genes}
\usage{
plotHeatmapDE(
  object,
  of_sample = "",
  data,
  across,
  across_subset = NULL,
  n_barcode_spots = 100,
  verbose = TRUE,
  hm_colors = viridis::viridis(15),
  ...
)
}
\arguments{
\item{data}{A data.frame containing at least the character or factor variables
\emph{cluster} and \emph{gene}.}

\item{across}{Character value. Denotes feature-variable used to perform the differentially
gene expression analysis.}

\item{across_subset}{Character vector or NULL. The elements of the \emph{cluster}-variable
\code{data} you want to subset. If set to NULL no subsetting takes place.}

\item{n_barcode_spots}{The number of barcode-spots belonging to each cluster you want to
include in the matrix. Should be lower than the total number of barcode-spots of every cluster
and can be deployed in order to keep the heatmap clear and aesthetically pleasing.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{hm_colors}{A vector of colors to be used by the heatmap.}

\item{...}{Additional parameters given to \code{pheatmap::pheatmap}.}
}
\value{
A heatmap of type "pheatmap".
}
\description{
Takes the results from your differentially gene expression analysis
and uses the expression information of your spata-object to plot a heatmap displaying
the differentially expressed genes of every cluster.
}
