% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-pattern-recognition.R
\name{plotIntraPatternDistance}
\alias{plotIntraPatternDistance}
\title{Title}
\usage{
plotIntraPatternDistance(
  object,
  of_sample = "",
  of_pattern = "",
  clrp = "milo"
)
}
\arguments{
\item{of_hotspots}{}
}
\value{

}
\description{
Title
}
