% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-pattern-recognition.R
\name{plotIntraPatternDistance}
\alias{plotIntraPatternDistance}
\title{Plot pattern related}
\usage{
plotIntraPatternDistance(object, of_pattern = "", clrp = NULL, of_sample = NA)
}
\description{
Blueprint - currently not reliably working.
}
