% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-pattern-recognition.R
\name{plotPatternEnrichment}
\alias{plotPatternEnrichment}
\title{Plot gene set enrichment within found patterns}
\usage{
plotPatternEnrichment(
  object,
  of_hotspots = "",
  of_gene_sets = "",
  top_n = 10,
  clrp = NULL,
  with_ties = FALSE,
  of_sample = NA,
  ...
)
}
\description{
Blueprint - currently not reliably working.
}
