% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-pattern-recognition.R
\name{plotPrSummary}
\alias{plotPrSummary}
\title{Title}
\usage{
plotPrSummary(
  object,
  plot_type = "density_2d",
  display_points = NULL,
  pt_size = NULL,
  clrp = NULL,
  of_sample = NA
)
}
\arguments{
\item{clrp}{}
}
\value{

}
\description{
Title
}
