% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotScatter}
\alias{plotScatter}
\title{Title}
\usage{
plotScatter(
  object,
  variables,
  pt_size = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  verbose = NULL,
  of_sample = NA
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
