% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plot segmentation}
\usage{
plotSegmentation(object, of_sample, pt_size = 2, verbose = FALSE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample to be displayed specified as a character vector of
length one.}

\item{pt_size}{The size of the points specified as a numeric value.}

\item{verbose}{Logical value. If set to TRUE informative messages with respect
to the computational progress made will be printed.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the segmentation of a specified sample that was drawn with
\code{SPATA::createSegmentation()}.
}
