% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plot segmentation}
\usage{
plotSegmentation(object, of_sample, pt_size = 2)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.}

\item{pt_size}{The size of the points specified as a single numeric value.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the segmentation of a specified sample that was drawn with
\code{SPATA::createSegmentation()}.
}
