% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurface}
\alias{plotSurface}
\alias{plotSurface2}
\alias{plotSurfaceInteractive}
\title{Plot the sample}
\usage{
plotSurface(
  object,
  of_sample = "",
  color_to = NULL,
  method_gs = "mean",
  normalize = TRUE,
  smooth = TRUE,
  smooth_span = 0.02,
  pt_alpha = 1,
  pt_clr = "lightgrey",
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  pt_size = 2,
  display_image = FALSE,
  display_title = FALSE,
  assign = FALSE,
  assign_name,
  complete = FALSE,
  verbose = TRUE,
  ...
)

plotSurface2(
  coords_df,
  color_to,
  pt_size = 2,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  pt_clrp = "milo",
  image = NULL,
  ...
)

plotSurfaceInteractive(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{color_to}{The variable to be displayed by color:

 \itemize{
  \item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} as a character vector. If more than one gene is specified the average
  expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
  \item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
  }}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{pt_alpha}{The transparency of the points specified as single numeric value.}

\item{pt_clr}{The base color of every point displayed in the plot.}

\item{pt_clrp}{The color panel to be used if the specified variable displayed by
color is categorical/discrete. Run \code{all_colorpanels()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{all_colorspectra()} to see valid input..}

\item{pt_size}{The size of the points specified as a single numeric value.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}

\item{display_title}{Logical. If set to TRUE an informative title is displayed.}

\item{assign}{Logical. If set to TRUE a named list will be assigned to the global
environment. This list contains data and information to rebuild or additionally
customize the output plot of this function.}

\item{assign_name}{The name the assigned list is supposed to have specified as
a single character value.}

\item{complete}{Logical. If the provided spata-object has been subsetted by
\code{subsetBySegment()} the original sample is completed with grey barcode
spots.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{coords_df}{A data.frame containing information about every barcode-spot. Must contain the variables:
\describe{
 \item{\emph{barcodes}}{Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
 \item{\emph{sample}}{Character. The sample belonging of every barcode-spot.}
 \item{\emph{x}}{Numeric. The x-coordinates of every barcode.}
 \item{\emph{y}}{Numeric. The y-coordinates of every barcode.}
}}

\item{image}{An image of class \emph{Image} to be displayed in the background.
Easily accessible via \code{SPATA::image()}.}
}
\description{
Displays the spatial dimension of the sample and colors the
surface according to the expression of genes, gene sets or features.

\itemize{

 \item{ \code{plotSurface()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters.}
 \item{ \code{plotSurface2()} Takes a data.frame as the starting point.}
 \item{ \code{plotSurfaceInteractive()} Takes only the spata-object and opens a shiny
 application which allows for interactive plotting.}

}
}
