% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotSurface}
\alias{plotSurface}
\title{Plot surface}
\usage{
plotSurface(
  object,
  of_sample,
  color_to,
  smooth = F,
  smooth_span = 0.02,
  method_gs = "mean",
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  display_image = F,
  display_title = F,
  verbose = T
)
}
\arguments{
\item{object}{A valid object of class \emph{spata}.}

\item{of_sample}{The samples name specified as a character of length one.}

\item{color_to}{The information to be displayed by color specified as a
character vector. If you specify a feature or a gene set this vector needs
to be of length one. If you specify more than one gene the average
expression of these genes will be calculated.}

\item{smooth}{Logical value. If set to TRUE \code{plotSurface()} will smooth
the values displayed by color deploying \code{stats::loess()}.}

\item{smooth_span}{Numeric value, given to \code{stats::loess()} if
\code{smooth} is set to TRUE.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}. Ignored if \code{color_to} isn't a gene set.}

\item{pt_size}{The size of the points specified as a numeric value.}

\item{pt_alpha}{The transparency of the points specified as a numeric value.}

\item{pt_clrsp}{The colour spectrum used to display \code{color_to} if the
specified variable is continuous. Needs to be one of \emph{inferno, magma,
plasma, cividis or viridis}.}

\item{display_image}{Logical value to specify whether the histological image
of the sample is supposed to be displayed underneath the plot.

(Can be made visible with low \code{pt_alpha} or \code{color_to} set to NULL.)}

\item{display_title}{Logical value to specify whether a title is supposed to
be displayed.}

\item{verbose}{Logical value to specify whether informative messages are
supposed to be printed or not.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the sample and color the surface according to the
expression of genes and gene sets or other featured characteristics.
}
