% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurfaceComparison}
\alias{plotSurfaceComparison}
\alias{plotSurfaceComparison2}
\title{Plot several surface plots at the same time}
\usage{
plotSurfaceComparison(
  data,
  variables = NULL,
  pt_size = 2,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  image = NULL,
  ...
)

plotSurfaceComparison2(
  object,
  of_sample,
  variables,
  method_gs = "mean",
  smooth = FALSE,
  smooth_span = 0.02,
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  display_image = FALSE,
  assign = FALSE,
  assign_name,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A data.frame containing the numeric variables of interest.}

\item{variables}{The variables of interest specified as a character vector:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 \item{ \strong{Features}: Must be a numeric one of \code{getFeaturenNames()}}
 }}

\item{pt_size}{The size of the points specified as a single numeric value.}

\item{pt_alpha}{The transparency of the points specified as single numeric value.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable that is displayed by
color is continuous. Needs to be one of \emph{'inferno', 'magma', 'plasma', 'cividis' or 'viridis'}.}

\item{image}{An image of class \emph{Image} to be displayed in the background.
Easily accessible via \code{SPATA::image()}.}

\item{...}{Additional arguments given to \code{ggplot2::facet_wrap()}.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}

\item{assign}{Logical. If set to TRUE a named list will be assigned to the global
environment. This list contains data and information to rebuild or additionally
customize the output plot of this function.}

\item{assign_name}{The name the assigned list is supposed to have specified as
a single character value.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Plots a surface plot for every valid element in argument
\code{color_to} and colors it according to the element.

\itemize{
 \item{ \code{plotSurfaceComparison()} Takes a data.frame as the starting point. }
 \item{ \code{plotSurfaceComparison2()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters.}
 }
}
