% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotSurfaceComparison}
\alias{plotSurfaceComparison}
\title{Plot several surface plots at the same time}
\usage{
plotSurfaceComparison(
  object,
  of_sample,
  color_to,
  method_gs = "mean",
  smooth = FALSE,
  smooth_span = 0.02,
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  display_image = FALSE,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid object of class \emph{spata}.}

\item{of_sample}{The samples name specified as a character of length one.}

\item{color_to}{The genes and gene sets whoose expression you want to
compare specified as a character vector or a list.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}. Ignored if \code{color_to} isn't a gene set.}

\item{smooth}{Logical value. If set to TRUE \code{plotSurface()} will smooth
the values displayed by color deploying \code{stats::loess()}.}

\item{smooth_span}{Numeric value, given to \code{stats::loess()} if
\code{smooth} is set to TRUE.}

\item{pt_size}{The size of the points specified as a numeric value.}

\item{pt_alpha}{The transparency of the points specified as a numeric value.}

\item{pt_clrsp}{The colour spectrum used to display \code{color_to} if the
specified variable is continuous. Needs to be one of \emph{inferno, magma,
plasma, cividis or viridis}.}

\item{display_image}{Logical value to specify whether the histological image
of the sample is supposed to be displayed underneath the plot.
(Can be made visible with low \code{pt_alpha} or \code{color_to} set to NULL.)}

\item{assign}{Logical value. If set to TRUE a list of the data needed to plot the
returned plot will be assigned to the global environment.}

\item{assign_name}{The name of the list that is assigned to the global environment
specified as a character value that does not already exist.}

\item{verbose}{Logical value to specify whether informative messages are
supposed to be printed or not.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Plots a surface plot for every valid element in argument
\code{color_to} and colors the surface according to the expression of
genes and gene sets.
}
