% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurfaceHotspots}
\alias{plotSurfaceHotspots}
\title{Title}
\usage{
plotSurfaceHotspots(
  object,
  plot_type = "expression",
  clrsp = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_clrp = NULL,
  pt_size = NULL,
  pt_clrsp = NULL,
  smooth = NULL,
  smooth_span = NULL,
  display_points = NULL,
  display_image = NULL,
  display_labels = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
