% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurfaceHotspots}
\alias{plotSurfaceHotspots}
\title{Title}
\usage{
plotSurfaceHotspots(
  object,
  of_sample = "",
  plot_type = "expression",
  clrsp = "inferno",
  pt_alpha = 0.9,
  pt_clr = "lightgrey",
  pt_clrp = "milo",
  pt_size = 2,
  smooth = TRUE,
  smooth_span = 0.02,
  display_points = TRUE,
  display_image = FALSE,
  display_labels = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
