% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurfaceQuantiles}
\alias{plotSurfaceQuantiles}
\title{Title}
\usage{
plotSurfaceQuantiles(
  object,
  color_to,
  n_qntls = 5,
  keep_qntls = 1:n_qntls,
  pt_alpha = NULL,
  pt_clrp = NULL,
  pt_size = NULL,
  smooth = NULL,
  smooth_span = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{smooth_span}{}
}
\value{

}
\description{
Title
}
