% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plotTrajectory}
\alias{plotTrajectory}
\title{Plot trajectory}
\usage{
plotTrajectory(
  object,
  trajectory_name,
  of_sample,
  color_to = NULL,
  method_gs = "mean",
  smooth = FALSE,
  smooth_span = 0.02,
  pt_size = 2.5,
  pt_alpha = 1,
  pt_clr = "steelblue",
  pt_clrsp = "inferno",
  sgmt_size = 1,
  display_image = F,
  display_title = F,
  verbose = T
)
}
\arguments{
\item{object}{A valid object of class \emph{spata}.}

\item{trajectory_name}{The trajectory to plot specified as a character vector
of length one.}

\item{of_sample}{The samples name specified as a character of length one.}

\item{color_to}{The information to be displayed by color specified as a
character vector. If you specify a feature or a gene set this vector needs
to be of length one. If you specify more than one gene the average
expression of these genes will be calculated..}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}. Ignored if \code{color_to} isn't a gene set.}

\item{smooth}{Logical value. If set to TRUE \code{plotSurface()} will smooth
the values displayed by color deploying \code{stats::loess()}.}

\item{smooth_span}{Numeric value, given to \code{stats::loess()} if
\code{smooth} is set to TRUE.}

\item{pt_size}{The size of the points specified as a numeric value.}

\item{pt_alpha}{The transparency of the points specified as a numeric value.}

\item{pt_clr}{The color of the points if \code{color_to} is set to NULL.}

\item{pt_clrsp}{The color spectrum used to display \code{color_to} if the
specified variable is continuous. Needs to be one of \emph{inferno, magma,
plasma, cividis or viridis}.}

\item{sgmt_size}{The size of the segment arrrow specified as a numeric value.}

\item{display_image}{Logical value. If set to TRUE the image will be displayed
as the background. If set to FALSE barcodes that do not fall into the trajectory
will be displayed in grey.}

\item{display_title}{Logical value.}

\item{verbose}{Logical value. If set to TRUE informative messages with respect
to the computational progress made will be printed.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays a trajectory of a specified sample that was
drawn with \code{SPATA::createTrajectories()}.
}
