% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_family.R
\name{plotTrajectoryFeatures}
\alias{plotTrajectoryFeatures}
\alias{plotTrajectoryGenes}
\alias{plotTrajectoryGeneSets}
\title{Trajectory line plots}
\usage{
plotTrajectoryFeatures(
  object,
  trajectory_name,
  of_sample,
  features = "percent.mt",
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = T,
  verbose = TRUE
)

plotTrajectoryGenes(
  object,
  trajectory_name,
  of_sample,
  genes,
  average_genes = F,
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = T,
  verbose = T
)

plotTrajectoryGeneSets(
  object,
  trajectory_name,
  of_sample,
  gene_sets = "Neftel_NPC_Comb",
  method_gs = "mean",
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = T,
  verbose = T
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory to plot specified as a character vector
of length one.}

\item{of_sample}{The sample from which the trajectory derives specfified as a
character vector of length one.}

\item{features}{The features you want to display specified as a character
vector (only numeric features can be plotted with \code{plotTrajectoryFeatures()}).}

\item{smooth_method}{The smoothing method that will be used specified as a
character vector of length one (e.g. \emph{"lm", "glm", "gam", "loess"}).}

\item{smooth_span}{The amount of smoothing specified as a numeric vector of
length one given to \code{stats::loess()}.}

\item{smooth_se}{Logical value. If set to TRUE the confidence interval will be
displayed.}

\item{verbose}{Logical value. If set to TRUE informative messages with respect
to the computational progress made will be printed.}

\item{genes}{The genes to be displayed specified as a character vector.}

\item{average_genes}{Logical value. If set to TRUE coords_df will be joined with
the mean expression values of all genes specified.}

\item{gene_sets}{The gene sets to be displayed specified as a character
vector.}

\item{method_gs}{The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{mean} or one
of \emph{gsva, ssgsea, zscore, or plage}. The latter four will be given to
\code{gsva::GSVA()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays values along a trajectory direction.
}
