% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryFeatures}
\alias{plotTrajectoryFeatures}
\alias{plotTrajectoryGenes}
\alias{plotTrajectoryGeneSets}
\title{Trajectory line plots}
\usage{
plotTrajectoryFeatures(
  object,
  trajectory_name,
  features = NULL,
  smooth_method = NULL,
  smooth_se = NULL,
  smooth_span = NULL,
  binwidth = 5,
  clrp = NULL,
  display_trajectory_parts = NULL,
  split = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotTrajectoryGenes(
  object,
  trajectory_name,
  genes,
  average_genes = FALSE,
  binwidth = 5,
  clrp = NULL,
  smooth_method = NULL,
  smooth_se = NULL,
  smooth_span = NULL,
  display_trajectory_parts = NULL,
  split = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotTrajectoryGeneSets(
  object,
  trajectory_name,
  gene_sets,
  binwidth = 5,
  method_gs = NULL,
  smooth_method = NULL,
  smooth_span = NULL,
  smooth_se = NULL,
  clrp = NULL,
  display_trajectory_parts = NULL,
  split = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{features}{The features of interest specified as a character vector.}

\item{smooth_method}{The smoothing method that will be used specified as a
single character value (e.g. \emph{"lm", "glm", "gam", "loess"}).}

\item{smooth_se}{Logical. If set to TRUE the confidence interval will be
displayed.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}

\item{clrp}{Character value. The color panel to be used.
Run \code{all_colorpanels()} to see valid input.}

\item{display_trajectory_parts}{Logical. If set to TRUE the returned plot
visualizes the parts in which the trajectory has been partitioned while beeing
drawn.}

\item{split}{Logical. If set to TRUE sub plots for every specified gene, gene-set
or feature are displayed via \code{ggplot2::facet_wrap()}}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{Additional arguments given to \code{ggplot2::facet_wrap()} if argument
\code{split} is set to TRUE.}

\item{genes}{The genes of interest specified as a character vector.}

\item{average_genes}{Logical value. If set to TRUE the mean expression
of all specified genes will be calculated.}

\item{gene_sets}{The gene sets of interest specified as a character vector.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{discrete_feature}{Character value. The discrete feature of interest.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays values along a trajectory direction with
a smoothed lineplot.
}
