% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryFeaturesDiscrete}
\alias{plotTrajectoryFeaturesDiscrete}
\title{Trajectory barplots}
\usage{
plotTrajectoryFeaturesDiscrete(
  object,
  trajectory_name,
  of_sample = "",
  discrete_feature,
  clrp = "milo",
  display_trajectory_parts = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{discrete_feature}{Character value. The discrete feature of interest.}

\item{clrp}{Character value. The color panel to be used.
Run \code{all_colorpanels()} to see valid input.}

\item{display_trajectory_parts}{Logical. If set to TRUE the returned plot
visualizes the parts in which the trajectory has been partitioned while beeing
drawn.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{ggplot2::facet_wrap()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays discrete variables along a trajectory.
}
