% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryFit}
\alias{plotTrajectoryFit}
\alias{plotTrajectoryFitCustomized}
\title{Display trajectory fit}
\usage{
plotTrajectoryFit(
  object,
  trajectory_name,
  of_sample = "",
  variable,
  method_gs = "mean",
  binwidth = 5,
  smooth = TRUE,
  smooth_span = 0.2,
  display_residuals = FALSE,
  verbose = TRUE,
  ...
)

plotTrajectoryFitCustomized(
  object,
  trajectory_name,
  of_sample = "",
  variable,
  customized_trends,
  method_gs = "mean",
  binwidth = 5,
  smooth = TRUE,
  smooth_span = 0.2,
  display_residuals = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{variable}{The gene or gene set of interest specified as a character value.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}

\item{display_residuals}{Logical. If set to TRUE the residuals are displayed
via a red line.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional parameters given to \code{ggplot2::facet_wrap()}.}

\item{customized_trends}{A data.frame or a named list. All numeric variables are considered to correspond to customized trends
the trajectory of interest might adopt. The names of the respective variables will correspond to the name
with which you want to refer to the trend later on.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the trend of a trajectory in comparison to a variety
of models / mathematical curves.
}
