% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryFit}
\alias{plotTrajectoryFit}
\title{Display trajectory fit}
\usage{
plotTrajectoryFit(rtdf, variable, display_residuals = FALSE, ...)
}
\arguments{
\item{rtdf}{A ranked trajectory data.frame.}

\item{display_residuals}{Logical. If set to TRUE the residuals are displayed
via a ret line.}

\item{...}{Additional parameters given to \code{ggplot2::facet_wrap()}.}

\item{gene_set}{The gene set of interest specified as a character value.}

\item{gene}{The gene of interest specified as a character value.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displayes the trend of a trajectory in comparison to a variety
of models / mathematical curves and thus how well it fits to each of them.
}
