% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryFit}
\alias{plotTrajectoryFit}
\title{Display trajectory fit}
\usage{
plotTrajectoryFit(object, trajectory_name, of_sample = "", variable,
  method_gs = "mean", accuracy = 5, display_residuals = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{variable}{The gene or gene set of interest specified as a character value.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{accuracy}{Numeric. Given to \code{accuracy}-argument of
\code{plyr::round_any()}. Determines how many barcode-spots will be summarized
as one sub-trajectory-part.}

\item{display_residuals}{Logical. If set to TRUE the residuals are displayed
via a red line.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional parameters given to \code{ggplot2::facet_wrap()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the trend of a trajectory in comparison to a variety
of models / mathematical curves.
}
