% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotUmap}
\alias{plotUmap}
\alias{plotTsne}
\alias{plotPca}
\title{Plot dimensional reduction}
\usage{
plotUmap(
  object,
  of_sample = "",
  color_to = NULL,
  method_gs = "mean",
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  pt_clrp = "milo",
  pt_clr = "black",
  normalize = TRUE,
  verbose = TRUE
)

plotTsne(
  object,
  of_sample = "",
  color_to = NULL,
  method_gs = "mean",
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  pt_clrp = "milo",
  normalize = TRUE,
  verbose = TRUE
)

plotPca(
  object,
  of_sample = "",
  color_to = NULL,
  n_pcs = 10,
  method_gs = "mean",
  pt_size = 1.5,
  pt_alpha = 0.9,
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  pt_clr = "black",
  normalize = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{color_to}{The variable to be displayed by color:

 \itemize{
  \item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} as a character vector. If more than one gene is specified the average
  expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
  \item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
  }}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{pt_size}{The size of the points specified as a single numeric value.}

\item{pt_alpha}{The transparency of the points specified as single numeric value.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{all_colorspectra()} to see valid input..}

\item{pt_clrp}{The color panel to be used if the specified variable displayed by
color is categorical/discrete. Run \code{all_colorpanels()} to see valid input.}

\item{pt_clr}{The base color of every point displayed in the plot.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{n_pcs}{Numeric value. Determines the number of principal components to be plotted.
Must be an even number.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the dimensional reduction and maps gene, gene-set
or feature information onto the color-aesthetic.
}
