% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-analysis.R
\name{rankTrajectoryGeneSets}
\alias{rankTrajectoryGeneSets}
\alias{rankTrajectoryGenes}
\title{Trajectory expression analysis}
\usage{
rankTrajectoryGeneSets(stdf)

rankTrajectoryGenes(stdf)
}
\arguments{
\item{stdf}{A summarized trajectory data.frame. (e.g. obtained by
\code{getSummarizedTrajectoryDf()}).}
}
\value{
A ranked trajectory data.frame. These nested data.frames contain
the following variables:

A nested data.frame with information about the dynamics of each gene
or gene set.
}
\description{
These functions analyze the expression dynamics along
a specified trajectory by fitting a variety of models to the trajectory's
expression trend.
}
