% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-recognition-main.R
\name{runPatternRecognition}
\alias{runPatternRecognition}
\title{Title}
\usage{
runPatternRecognition(
  object,
  genes = 1500,
  genes_additional = NULL,
  with_ties = TRUE,
  method_pr = "hotspot",
  threshold_qntl = 0.8,
  threshold_stw = 0.5,
  threshold_stpv = 0.1,
  smooth = TRUE,
  smooth_span = 0.02,
  verbose = TRUE,
  of_sample = NA
)
}
\arguments{
\item{n_start}{}
}
\value{

}
\description{
Title
}
