% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-input-options.R
\name{validActivationFunctions}
\alias{validActivationFunctions}
\alias{validHierarchicalClusterMethods}
\alias{validDeAnalysisMethods}
\alias{validDefaultDirectories}
\alias{validDefaultInstructionSlots}
\alias{validDimRedMethods}
\alias{validColorPanels}
\alias{validPatternRecognitionMethods}
\alias{validColorSpectra}
\title{Obtain valid argument inputs}
\usage{
validActivationFunctions()

validHierarchicalClusterMethods()

validDeAnalysisMethods()

validDefaultDirectories()

validDefaultInstructionSlots()

validDimRedMethods()

validColorPanels()

validPatternRecognitionMethods()

validColorSpectra()
}
\value{
Character vectors or named lists of such.
}
\description{
These function simply return valid input options
for recurring arguments.
}
